% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs--data.R
\docType{data}
\name{albufera_clusters}
\alias{albufera_clusters}
\alias{albufera_cluster_geometries}
\title{Albufera Rice Paddy Clusters}
\format{
\subsection{\code{albufera_clusters}}{

A data frame with one row per cluster, and the following columns:
\describe{
\item{cluster_id}{character. Unique identifier of cluster.}
\item{ditch}{character. Ditch to which the cluster pertains.}
\item{area}{numeric. Area (in squared meters) of the cluster.}
\item{tancat}{logical. Whether the cluster is a "tancat".}
\item{rice_variety}{character. Variety of rice planted in the cluster.}
\item{case}{factor. ...?}
}
}

\subsection{\code{albufera_cluster_geometries}}{

A data frame with one row per cluster, and the columns:
\describe{
\item{cluster_id}{character. Unique identifier of cluster.}
\item{geometry}{vector of class \code{sfc_MULTIPOLYGON}.}
}
}
}
\source{
Martínez-Megías, C., Mentzel, S., Fuentes-Edfuf, Y., Moe, S. J., &
Rico, A. (2023). Influence of climate change and pesticide use practices on
the ecological risks of pesticides in a protected Mediterranean wetland: A
Bayesian network approach. \emph{Science of The Total Environment, 878},
163018. \url{https://doi.org/10.1016/j.scitotenv.2023.163018}
}
\usage{
albufera_clusters

albufera_cluster_geometries
}
\description{
Datasets containing information on rice field clusters into which the
rice production area of the Albufera Natural Park is decomposed. For the
actual definition of these clusters, see
\href{https://doi.org/10.1016/j.scitotenv.2023.163018}{Martínez-Megías et al. (2023)}.
}
\details{
Both datasets have the same cardinality (one row per cluster), and they
are separated merely for efficiency reasons. They can be joined using the
common \code{cluster_id} key column.
}
\keyword{datasets}
