% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{albufera_management_chemicals}
\alias{albufera_management_chemicals}
\title{Albufera Chemicals Management Data}
\format{
\subsection{\code{albufera_management_chemicals}}{

The dataset has one row per chemical scheduled application to a given rice
variety, and the following columns:
\describe{
\item{day}{numeric. Scheduled day, counted starting from the sowing day,
for the application under consideration.}
\item{variety}{character. Rice variety to which the chemical is applied.}
\item{chemical}{character. Name of applied chemical.}
\item{amount}{numeric. Amount of chemical applied.}
}
}
}
\source{
TODO #68
}
\usage{
albufera_management_chemicals
}
\description{
Data on application schedules of chemicals used in rice paddies in the
Albufera National Park.
}
\keyword{datasets}
