% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{albufera_petp}
\alias{albufera_petp}
\title{Albufera Precipitation and Evapotranspiration Data}
\format{
\subsection{\code{albufera_petp}}{

A data frame with 6140 rows and 5 columns:
\describe{
\item{date}{??}
\item{P}{??}
\item{ETP}{??}
}
}
}
\source{
???
}
\usage{
albufera_petp
}
\description{
Albufera Precipitation and Evapotranspiration Data
}
\keyword{datasets}
