% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_balance.R
\name{albufera_storage_curve}
\alias{albufera_storage_curve}
\title{Albufera Lake storage curve}
\usage{
albufera_storage_curve(
  level,
  intercept = 16.7459 * 1e+06,
  slope = 23.6577 * 1e+06
)
}
\arguments{
\item{level}{numeric vector. Levels of the lake (in meters above sea level).
Accepts vectorized input.}

\item{intercept}{a number. Intercept of the linear storage curve.}

\item{slope}{a number. Slope of the linear storage curve}
}
\value{
A numeric vector of the same length of \code{level}. Computed volumes (in
\eqn{\text{m}^3}).
}
\description{
Computes the Albufera lake's volume from level using a linear storage curve,
\emph{i.e.} \eqn{\text{Volume} = \text{slope} \times \text{Level} + \text{intercept}}.
}
\details{
The default values are taken from the CHJ report
\href{https://www.chj.es/Descargas/ProyectosOPH/Consulta\%20publica/PHC-2015-2021/ReferenciasBibliograficas/HumedalesZonasProtegidas/CHJ,2012.Aquatool_Albufera.pdf}{Modelo de seguimiento de l’Albufera de Valencia con AQUATOOLDMA.}
}
\author{
Pablo Amador Crespo, Valerio Gherardi
}
