% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs--data.R
\docType{data}
\name{albufera_weather}
\alias{albufera_weather}
\title{Albufera Weather Daily Data}
\format{
\subsection{\code{albufera_weather}}{

A data frame with 8552 rows and 3 columns:
\describe{
\item{date}{Date of measurement}
\item{temperature_ave}{Average temperature (TODO #50).}
\item{temperature_min}{Minimum temperature.}
\item{temperature_max}{Maximum temperature.}
\item{precipitation_mm}{Daily precipitation in millimiters.}
\item{evapotranspiration_mm}{Daily evapotranspiration in millimiters.}
}
}
}
\source{
TODO #50
}
\usage{
albufera_weather
}
\description{
TODO #50
}
\seealso{
Other input data: 
\code{\link{albufera_outflows}}
}
\concept{input data}
\keyword{datasets}
