% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca--main.R
\name{ca}
\alias{ca}
\alias{setup_ca}
\title{CA: Chemical Applications}
\usage{
setup_ca(simulation, ca_schedules_df = erahumed::albufera_ca_schedules)
}
\arguments{
\item{simulation}{An object of class \link{erahumed_simulation}.}

\item{ca_schedules_df}{a \code{data.frame} following the template of
\link{albufera_ca_schedules}. Each row of this data.frame corresponds to a
scheduled application. The semantics of columns are the same as in
\link{albufera_ca_schedules}.}
}
\value{
An objects of class \link{erahumed_simulation}.
}
\description{
This simulation layer simulates the application of chemicals to rice paddy
clusters, based on a previously computed simulation of local hydrological
balance. The result is a set of time series of applied doses, one for each
applied chemical.
}
\details{
The output \code{data.frame} extends the output of the underlying \link{hbp}
layer, preserving its cardinality (one row per cluster per day). The
additional columns, named as the chemicals appearing in \code{ca_schedules_df},
provide the time series of applied doses, expressed in kilograms.
}
\seealso{
Other simulation layers: 
\code{\link{ct}},
\code{\link{setup_hba}()},
\code{\link{setup_hbp}()},
\code{\link{setup_inp}()}
}
\author{
Pablo Amador Crespo, Valerio Gherardi
}
\concept{simulation layers}
