% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux--chemicals.R
\name{chemicals}
\alias{chemicals}
\alias{chemical_properties}
\title{Information on supported chemicals}
\usage{
chemicals()

chemical_properties(name)
}
\arguments{
\item{name}{A string. Short name of the chemical for which properties are
required, one of the entries of the vector returned by \code{chemicals()}.}
}
\value{
A character vector for \code{chemicals()}, a list for \code{chemical_properties()}.
}
\description{
These helpers provide information on the set of chemicals supported by
ERAHUMED: \code{chemicals()} lists the short names of supported chemicals,
whereas \code{chemical_properties()} returns the properties of a given chemical.
}
