% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfcm--presets.R, R/rfcm--s3-class.R
\name{cluster_map}
\alias{cluster_map}
\alias{default_cluster_map}
\alias{new_cluster_map}
\title{Cluster-Based Management Assignment}
\usage{
default_cluster_map(seed = 840)

new_cluster_map(default_management_system = new_management_system())
}
\arguments{
\item{seed}{\verb{[numeric(1)]} \cr
Seed for random number generation in the assignation of clusters to
management systems.}

\item{default_management_system}{\code{[}\link[=management_system]{erahumed_management_system}\verb{]} \cr
A default management system assigned to all clusters initially. Typically
created with \code{\link[=new_management_system]{new_management_system()}} or a helper like \code{\link[=jsendra]{jsendra()}}.}
}
\value{
An object of class \code{erahumed_cluster_map}.
}
\description{
These functions define how different rice field management systems are
assigned to spatial clusters within the simulation. Each cluster can be
associated with a specific management strategy, allowing for heterogeneous
management across the simulation domain.
}
\details{
\itemize{
\item \code{new_cluster_map()} initializes a cluster map with all clusters assigned
the same default management system.
\item \code{default_cluster_map()} provides a predefined map inspired by current
practices in the Albufera Natural Park. Specifically, it uses the \emph{J. Sendra}
system as the default, and allocates small proportions of \emph{Bomba} (10\% in
tancats) and \emph{Clearfield} (10\% in ditches 1–19). This map is a convenient
starting point for scenario simulation or customization.
}

Cluster assignments can be modified using \code{\link[=allocate_surface]{allocate_surface()}}.
}
\seealso{
\code{\link[=new_management_system]{new_management_system()}}, \code{\link[=allocate_surface]{allocate_surface()}}
}
