% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_output.R
\name{component_output}
\alias{component_output}
\title{Extract model component outputs}
\usage{
component_output(object, ...)
}
\arguments{
\item{component}{An object inheriting from class
\link{erahumed_model_component}.}
}
\value{
A \code{data.frame}.
}
\description{
Extracts the output \code{data.frame} of a \link{erahumed_model_component}.
}
\author{
Valerio Gherardi
}
