% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta--component-parameters.R
\name{component_parameters}
\alias{component_parameters}
\alias{component_parameters.erahumed_model_component}
\alias{component_parameters.erahumed_model}
\title{Extract model component parameters}
\usage{
component_parameters(object, ...)

\method{component_parameters}{erahumed_model_component}(object, ...)

\method{component_parameters}{erahumed_model}(object, component, ...)
}
\arguments{
\item{object}{The object from which the component parameters are to be
extracted from (see above for available methods).}

\item{...}{further arguments passed to or from other methods.}

\item{component}{A string. Name of the component whose parameters \code{list}
is to be extracted.}
}
\value{
A \code{list}.
}
\description{
Extracts the parameters used in the computation of a model component.
}
\author{
Valerio Gherardi
}
