% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hb_utils.R
\name{compute_ditch_inflow_pct}
\alias{compute_ditch_inflow_pct}
\title{Compute Ditch Inflow Percent}
\usage{
compute_ditch_inflow_pct(ditch, area)
}
\arguments{
\item{ditch, area}{vectors of equal lengths, of type \code{"character"} and
\code{"double"}, respectively. The i-th entries of these vectors correspond to the
i-th cluster of rice paddies, and inform the ditch to which this cluster
belongs to, and the area of the cluster.}
}
\value{
A dataframe containing two columns, \code{ditch} and \code{inflow_pct},
specifying a ditch and the amount of inflow corresponding to it,
respectively.
}
\description{
Computes the fraction of water that flows through each ditch, as
the fraction of total surface covered by the clusters adjacent to said ditch.
}
\details{
TODO: other options to obtain ditch percents (e.g. empirical data by Soria et al.)?
}
\author{
Pablo Amador Crespo, Valerio Gherardi
}
