% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctc--main.R
\name{ctc}
\alias{ctc}
\alias{setup_ctc}
\title{Chemical Transport}
\usage{
setup_ctc(
  simulation,
  drift,
  covmax,
  jgrow,
  SNK,
  dact_m,
  css_ppm,
  foc,
  bd_g_cm3,
  qseep_m_day,
  wilting,
  fc
)
}
\arguments{
\item{simulation}{\code{[}\link{erahumed_simulation}\verb{]} \cr
The simulation object being modified.}

\item{drift}{\verb{[numeric(1)]} \cr A number between \code{0} and \code{1}. Percentage of chemical applications lost to drift.}

\item{covmax}{\verb{[numeric(1)]} \cr A number between \code{0} and \code{1}. Interception potential of foliage at crop maturation.}

\item{jgrow}{\verb{[numeric(1)]} \cr A positive integer. Length (in days) of crop maturation cycle.}

\item{SNK}{\verb{[numeric(1)]} \cr A number between \code{0} and \code{1}. (TODO).}

\item{dact_m}{\verb{[numeric(1)]} \cr A positive number. Active sediment layer depth, expressed in meters.}

\item{css_ppm}{\verb{[numeric(1)]} \cr A positive number. Suspended sediment concentration, expressed in parts per million.}

\item{foc}{\verb{[numeric(1)]} \cr A number between \code{0} and \code{1}. Fraction of organic content.}

\item{bd_g_cm3}{\verb{[numeric(1)]} \cr A positive number. Bulk density of the sediment, expressed in grams per cubic centimeter.}

\item{qseep_m_day}{\verb{[numeric(1)]} \cr A positive number. Seepage rate, expressed in meters per day.}

\item{wilting}{\verb{[numeric(1)]} \cr A number between \code{0} and \code{1}. Wilting point.}

\item{fc}{\verb{[numeric(1)]} \cr A number between \code{0} and \code{1}. Field capacity.}
}
\value{
An object of class \link{erahumed_simulation}.
}
\description{
This layer computes the temporal dynamics of chemicals' masses in rice fields, in the three compartments of foliage, water and sediment.
}
\seealso{
Other simulation layers: 
\code{\link{ca}}
}
\concept{simulation layers}
