% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ditch_inflow_pcts}
\alias{ditch_inflow_pcts}
\title{Ditch Inflow Percents}
\format{
\subsection{\code{ditch_inflow_pcts}}{

A dataframe with one row per month, and columns:
\describe{
\item{month}{Integer. Month number (1 = January, 2 = February, \emph{etc.}).}
\item{dXX}{Double. Percent of total inflow contributed by ditch number XX.}
}
}
}
\source{
\url{https://www.mdpi.com/2306-5338/8/1/37}
}
\usage{
ditch_inflow_pcts
}
\description{
Percents of total inflows to the Albufera Lake from individual ditches, by
month. Based on observational data.
}
\keyword{datasets}
