% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erahumed-package.R
\docType{package}
\name{erahumed-package}
\alias{erahumed}
\alias{erahumed-package}
\title{erahumed: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vgherard/erahumed}
  \item Report bugs at \url{https://github.com/vgherard/erahumed/issues}
}

}
\author{
\strong{Maintainer}: Valerio Gherardi \email{vgherard840@gmail.com} (\href{https://orcid.org/0000-0002-8215-3013}{ORCID})

}
\keyword{internal}
