% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs--parse-yml-documentation.R
\name{erahumed_docs}
\alias{erahumed_docs}
\title{Parameters documentation}
\usage{
erahumed_docs(...)
}
\arguments{
\item{param}{string. Parameter name.}
}
\value{
A list.
}
\description{
Extracts the documentation for a specific ERAHUMED parameter as an R string.
For internal use only.
}
