% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta--layers.R
\name{erahumed_layers}
\alias{erahumed_layers}
\title{ERAHUMED simulation layers}
\usage{
erahumed_layers()
}
\value{
A character vector.
}
\description{
Enumerates the ERAHUMED simulation layers
(\emph{cf.} \link{erahumed_simulation_interface}), in dependency order (from
upstream to downstream). A schematic view is available
\href{https://erahumed.github.io/erahumed/articles/pipeline-scheme.html}{on the package website}.
}
\details{
For further information on the various layers, see \verb{?<layer_name>}.
}
