% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erahumed_model.R
\name{erahumed_model}
\alias{erahumed_model}
\title{Initialize an ERAHUMED model}
\usage{
erahumed_model()
}
\value{
An object of class \code{erahumed_model}, that represents a blank ERAHUMED
model (with no model component computed yet).
}
\description{
Initializes an ERAHUMED model. This is the starting point for the ERAHUMED
model chain, whose computations are performed by \verb{compute_*()} functions.
}
\examples{
m <- erahumed_model()
m

}
\author{
Valerio Gherardi
}
