% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs--parameter-descriptions.R
\name{erahumed_param_docs}
\alias{erahumed_param_docs}
\title{Parameters documentation}
\usage{
erahumed_param_docs(param)
}
\arguments{
\item{param}{string. Parameter name.}
}
\value{
A list.
}
\description{
Extracts the documentation for a specific ERAHUMED parameter as an R string.
For internal use only.
}
