% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta--simulation.R
\name{erahumed_simulation}
\alias{erahumed_simulation}
\title{Initialize an ERAHUMED simulation}
\usage{
erahumed_simulation(
  date_start = "2023-01-01",
  date_end = "2024-12-31",
  seed = 840,
  outflows_df = erahumed::albufera_outflows,
  storage_curve_slope_m2 = 23.66 * 1e+06,
  storage_curve_intercept_m3 = 16.75 * 1e+06,
  petp_surface_m2 = 53.9 * 1e+06,
  weather_df = erahumed::albufera_weather,
  ideal_flow_rate_cm = 4.27,
  height_thresh_cm = 0.5,
  ditch_level_m = 1,
  covmax = 0.5,
  jgrow = 152,
  dact_m = 0.036,
  css_ppm = 50,
  foc_ss = 0.1,
  foc_sed = 0.082,
  bd_g_cm3 = 1.95,
  porosity = 0.11,
  ksetl_m_day = 2,
  rfms_map = default_rfms_map(seed = seed),
  .progress = message
)
}
\arguments{
\item{date_start}{\verb{[Date(1)]} \cr A single-value \code{Date} vector (or a convertible object). Defines the starting date of the date interval for the simulation.}

\item{date_end}{\verb{[Date(1)]} \cr A single-value \code{Date} vector (or a convertible object). Defines the ending date of the date interval for the simulation.}

\item{seed}{\verb{[numeric(1)]} \cr Seed for random number generation used by the simulation algorithms.}

\item{outflows_df}{\verb{[data.frame]} \cr Time-series dataset that provides the observational hydrological data on the Albufera lake, along the template of \link{albufera_outflows} (the default value).}

\item{storage_curve_slope_m2}{\verb{[numeric(1)]} \cr Slope of the (linear) storage curve of the Albufera Lake, in square meters. Multiplying this value by the change in the lake's water level gives the corresponding change in water \emph{volume}. The default numeric values are derived from the CHJ report \href{https://www.chj.es/Descargas/ProyectosOPH/Consulta\%20publica/PHC-2015-2021/ReferenciasBibliograficas/HumedalesZonasProtegidas/CHJ,2012.Aquatool_Albufera.pdf}{\emph{Modelo de seguimiento de l’Albufera de Valencia con AQUATOOLDMA}}.}

\item{storage_curve_intercept_m3}{\verb{[numeric(1)]} \cr Intercept of the linear storage curve for the Albufera Lake, expressed in cubic meters. This value represents the lake's water volume when the water level is at sea level. The default numeric values are derived from the CHJ report \href{https://www.chj.es/Descargas/ProyectosOPH/Consulta\%20publica/PHC-2015-2021/ReferenciasBibliograficas/HumedalesZonasProtegidas/CHJ,2012.Aquatool_Albufera.pdf}{\emph{Modelo de seguimiento de l’Albufera de Valencia con AQUATOOLDMA}}.}

\item{petp_surface_m2}{\verb{[numeric(1)]} \cr The surface area (in square meters) used to convert precipitation and evapotranspiration per unit area into volume changes for the Albufera Lake. The default numeric value was computed by the package authors and corresponds to the average flooded surface of the Albufera lake and its irrigation ditches.}

\item{weather_df}{\verb{[data.frame]} \cr A dataset that provides the relevant metereological time series, along the template of \link{albufera_weather} (the default value).}

\item{ideal_flow_rate_cm}{\verb{[numeric(1)]} \cr Ideal inflow/outflow of a cluster, for days in which the cluster is scheduled to be in flux (\emph{i.e.} when being simultaneously irrigated and drained). Expressed in centimeters of water depth per day.}

\item{height_thresh_cm}{\verb{[numeric(1)]} \cr A positive number. Height threshold for water levels, below which a cluster is considered to be emptied.}

\item{ditch_level_m}{\verb{[numeric(1)]} \cr Constant water depth in ditches.}

\item{covmax}{\verb{[numeric(1)]} \cr A number between \code{0} and \code{1}. Maximum potential fraction of applied chemicals intercepted by foliage at crop maturation.}

\item{jgrow}{\verb{[numeric(1)]} \cr A positive integer. Length (in days) of crop maturation cycle in rice paddies.}

\item{dact_m}{\verb{[numeric(1)]} \cr A positive number. Active sediment layer depth, where "active" refers to the portion of the sediment that is actually involved in chemical exchange or transport processes. Expressed in meters.}

\item{css_ppm}{\verb{[numeric(1)]} \cr A positive number. Concentration of water suspended solids, expressed in parts per million.}

\item{foc_ss}{\verb{[numeric(1)]} \cr A number between \code{0} and \code{1}. Fraction of organic matter content within suspended solids.}

\item{foc_sed}{\verb{[numeric(1)]} \cr A number between \code{0} and \code{1}. Fraction of organic matter content within sediment.}

\item{bd_g_cm3}{\verb{[numeric(1)]} \cr A positive number. Bulk density of the sediment (including both the solid material and the pore spaces). Expressed in grams per cubic centimeter.}

\item{porosity}{\verb{[numeric(1)]} \cr A number between \code{0} and \code{1}. The fraction of the total sediment volume that is occupied by pore spaces.}

\item{ksetl_m_day}{\verb{[numeric(1)]} \cr Settling rate, representing particle-bound transfer to sediment.}

\item{rfms_map}{\verb{[erahumed_rfms_map]} \cr An object of class \link{rfms_map} that defines how different rice field management systems (RFMSs) are assigned to spatial clusters within the simulation. This object also encapsulates the full set of user-defined agrochemical configurations, including custom chemicals and RFMSs, and serves as the main interface for scenario customization. See \link{rfms} for an overview of RFMS definitions and customization options.}

\item{.progress}{A function used to report simulation progress.
It should accept a single character string as input, representing the
current stage of the simulation (e.g., \code{"Computing hydrology: lake"}).
This parameter is primarily intended for user feedback mechanisms (e.g.,
console messages). By default, it prints a message to the R console.}
}
\value{
An object of class \code{erahumed_simulation}.
}
\description{
Initializes an ERAHUMED simulation.
See the \href{https://erahumed.github.io/erahumed/articles/erahumed-workflow.html}{main package vignette}
for a practical, step-by-step overview of the simulation workflow.
For a detailed explanation of the underlying models, algorithms, and
assumptions, refer to the
\href{https://erahumed.github.io/erahumed-book/}{user manual}.
}
\details{
The \code{rfms_map} argument plays a central role in the customization
capabilities of the ERAHUMED model. Beyond mapping rice field management
systems (RFMSs) to spatial clusters, it also carries the full definitions of
custom chemicals and RFMSs configured by the user. This makes it the primary
interface for building and running alternative scenarios involving changes in
pesticide use or management strategies.
For a detailed explanation of RFMS concepts, structure, and configuration,
see the \href{https://erahumed.github.io/erahumed-book/chapters/rfms.html}{RFMS section of the user manual}.
}
\examples{
erahumed_simulation()

}
