% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R, R/risk.R
\name{exposure}
\alias{exposure}
\alias{setup_exposure}
\alias{compute_exposure}
\alias{extract_exposure}
\alias{extract_risk}
\title{Pesticide exposure in the Albufera Natural Park}
\usage{
setup_exposure(
  simulation,
  ca_schedules_df = erahumed::albufera_ca_schedules,
  drift = 0,
  covmax = 0.5,
  jgrow = 152,
  SNK = 0,
  dact_m = 0.1,
  css_ppm = 50,
  foc = 0.17,
  bd_g_cm3 = 1.5,
  qseep_m_day = 0,
  wilting = 0.24,
  fc = 0.35
)

compute_exposure(simulation)

extract_exposure(simulation, element = c("lake", "ditch", "cluster"))

extract_risk(simulation, element = c("lake", "ditch", "cluster"))
}
\arguments{
\item{simulation}{\code{[}\link{erahumed_simulation}\verb{]} \cr
The simulation object being modified.}

\item{drift}{\verb{[numeric(1)]} \cr A number between \code{0} and \code{1}. Percentage of chemical applications lost to drift.}

\item{covmax}{\verb{[numeric(1)]} \cr A number between \code{0} and \code{1}. Interception potential of foliage at crop maturation.}

\item{jgrow}{\verb{[numeric(1)]} \cr A positive integer. Length (in days) of crop maturation cycle.}

\item{SNK}{\verb{[numeric(1)]} \cr A number between \code{0} and \code{1}. (TODO).}

\item{dact_m}{\verb{[numeric(1)]} \cr A positive number. Active sediment layer depth, expressed in meters.}

\item{css_ppm}{\verb{[numeric(1)]} \cr A positive number. Suspended sediment concentration, expressed in parts per million.}

\item{foc}{\verb{[numeric(1)]} \cr A number between \code{0} and \code{1}. Fraction of organic content.}

\item{bd_g_cm3}{\verb{[numeric(1)]} \cr A positive number. Bulk density of the sediment, expressed in grams per cubic centimeter.}

\item{qseep_m_day}{\verb{[numeric(1)]} \cr A positive number. Seepage rate, expressed in meters per day.}

\item{wilting}{\verb{[numeric(1)]} \cr A number between \code{0} and \code{1}. Wilting point.}

\item{fc}{\verb{[numeric(1)]} \cr A number between \code{0} and \code{1}. Field capacity.}

\item{element}{\verb{[character(1)]} \cr
String specifying the landscape element for which simulation results are
requested. Either \code{"lake"}, \code{"ditch"}, or \code{"cluster"}.}
}
\value{
Functions \verb{setup_*()} and \verb{compute_*()} return a
\link{erahumed_simulation}, while \verb{extract_*()} returns a \code{data.frame}.
}
\description{
These functions are used to setup, run and extract simulation
results for pesticide exposure in the Albufera Natural Park system.
}
