% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta--layer-extractors.R
\name{get_layer_output}
\alias{get_layer_output}
\alias{get_layer_output.erahumed_simulation_layer}
\alias{get_layer_output.erahumed_simulation}
\title{Extract simulation layer outputs}
\usage{
get_layer_output(object, ...)

\method{get_layer_output}{erahumed_simulation_layer}(object, ...)

\method{get_layer_output}{erahumed_simulation}(object, layer, ...)
}
\arguments{
\item{object}{The object from which the layer output is to be extracted
from (see above for available methods).}

\item{...}{further arguments passed to or from other methods.}

\item{layer}{A string. Name of the layer whose output \code{data.frame}
is to be extracted.}
}
\value{
A \code{data.frame}.
}
\description{
Extracts the output \code{data.frame} of a \link{erahumed_simulation_layer}.
}
\author{
Valerio Gherardi
}
