% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta--extractors.R
\name{get_results}
\alias{get_results}
\title{Extract simulation results}
\usage{
get_results(
  simulation,
  component = c("hydrology", "exposure", "risk"),
  element = c("lake", "ditch", "cluster")
)
}
\arguments{
\item{simulation}{\code{[}\link{erahumed_simulation}\verb{]} \cr
The simulation object being modified.}

\item{component}{\verb{[character(1)]} \cr
The simulation component to be extracted.
Either \code{"hydrology"}, \code{"exposure"}, or \code{"risk"}.}

\item{element}{\verb{[character(1)]} \cr
The landscape element for which simulation results are requested.
Either \code{"lake"}, \code{"ditch"}, or \code{"cluster"}.}
}
\value{
A \code{data.frame}.
}
\description{
Extract ERAHUMED simulation results from an
\link{erahumed_simulation} object.
}
