% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hb_global.R
\name{hb_global}
\alias{hb_global}
\title{Global Hydrological Balance}
\usage{
hb_global(
  level,
  rain_mm,
  evapotranspiration_mm,
  outflows,
  ...,
  storage_curve = erahumed::linear_storage_curve(slope = 1, intercept = 0),
  petp_surface = erahumed::linear_petp_surface(surface_P = 1, surface_ETP = 1)
)
}
\arguments{
\item{level}{numeric vector. Time series of lake levels, in meters.}

\item{rain_mm}{numeric vector. Time series of precipitation values, in millimiters.}

\item{evapotranspiration_mm}{numeric vector. Time series of evapotranspiration values, in
millimiters.}

\item{outflows}{a data.frame whose columns are the time series of outflows,
expressed in cube meters per second.}

\item{...}{additional columns to be appended in the returned data-frame. Each
of these additional (named) arguments should be a vector of the same length
implied by the previous arguments.}

\item{storage_curve}{a function that takes a numeric vector as input, and
returns a numeric vector of the same length. Function that converts lake
levels (passed through the \code{level} argument) into lake \emph{volumes}.}

\item{petp_surface}{a function that takes two numeric vectors of common
length as inputs, and returns a numeric vector of the same length. Function
that converts precipitation and evapotranspiration values (passed through
the \code{rain_mm} and \code{evapotranspiration_mm} arguments) into an overall volume change.}
}
\value{
An object of class \code{hb_global}, a lightweight wrapper of \code{data.frame}
with a few additional visualization methods (most prominently
\link{plot.hb_global}). The underlying data-frame contains as columns the
input time series, as well as the following calculated columns:
\itemize{
\item \code{volume} Volume time series, obtained from the storage curve.
\item \code{volume_change} Differenced time series of volume. The \eqn{n}-th is given
by \eqn{\Delta V _n \equiv V_{n+1}-V_n}, where \eqn{V_n} is volume at time
step \eqn{n}.
\item \code{petp_change} Time series. The amount \eqn{\Delta V _n ^\text{P-ETP}} of
volume change due to precipitation and evapotranspiration, obtained from the
P-ETP surface.
\item \code{outflow_total} Time series of total outflows, measured in cube meters per
second. This is the sum
\eqn{\sum _i O_i} of time series passed through the \code{outflows} argument,
plus an extra term \eqn{\delta O} of unaccounted outflow. The correction term
is defined by
\eqn{\delta O = \frac{\Delta V _n - \Delta V _n ^\text{P-ETP}}{24 \times 60 \times 60} -\sum _i O_i}
and is chosen in such a way to ensure that the total inflow is always
non-negative.
\item \code{outflow_extra} Time series. The unaccounted outflow term \eqn{\delta O}
described above.
\item \code{inflow_total} Time serie of total inflows, in cube meters per second.
This is computed as \eqn{I = \sum _{i} O_i + \delta O + \frac{\Delta V _n - \Delta V _n ^\text{P-ETP}}{24 \times 60 \times 60}}.
\item \code{residence_time_days}. Residence time, as modeled by \link{hbg_residence_time}.
}
}
\description{
Computes the global hydrological balance of a water basin from the
measurements of water level, outflows, and precipitation/evapotranspiration.
For analyzing Albufera data, you should not need to directly run this
function, and you can instead use the \link{albufera_hb_global} wrapper, that
calls \code{hb_global()} with the right arguments, extracted from the built-in
datasets.
}
