% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hb_local.R
\name{hb_local}
\alias{hb_local}
\title{Local Hydrological Balance}
\usage{
hb_local(
  date,
  cluster_id,
  ditch,
  ideal_height_cm,
  petp_cm,
  ideal_irrigation,
  ideal_draining,
  area_m2,
  total_inflow_lake,
  ...,
  ideal_flow_rate_cm = 5
)
}
\arguments{
\item{date}{vector. Can be either a character, formatted as YYYY-MM-DD, or
a \code{Date} object.}

\item{cluster_id}{character vector.}

\item{ditch}{character vector. Ditch to which the cluster specified by
\code{cluster_id} corresponds.}

\item{ideal_height_cm}{numeric vector. Ideal height of the cluster specified
by \code{cluster_id}, on the day specified by \code{date}.}

\item{petp_cm}{numeric vector. Precipitation minus evapotranspiration
(in cm), relevant for the \code{cluster_id}/\code{date} pair.}

\item{ideal_irrigation}{logical. Irrigation plan for the cluster specified
by \code{cluster_id}, on the day specified by \code{date} (is this cluster supposed to
be irrigated on this day?).}

\item{ideal_draining}{logical. Draining plan for the cluster specified
by \code{cluster_id}, on the day specified by \code{date} (is this cluster supposed to
be drained on this day?).}

\item{area_m2}{numeric vector. Area of the cluster specified by \code{cluster_id}.}

\item{total_inflow_lake}{numeric vector. Total inflow of lake (or,
equivalently, total outflow of clusters) on the day specified by \code{date}.}

\item{...}{additional columns to be appended in the returned data-frame. Each
of these additional (named) arguments should be a vector of the same length
implied by the previous arguments.}

\item{ideal_flow_rate_cm}{a positive number. Ideal inflow for days such
that \code{ideal_irrigation} and \code{ideal_draining} are both \code{TRUE}.}
}
\value{
An object of class \code{hb_local}, a lightweight wrapper of \code{data.frame}
with a few additional visualization methods (most prominently
\link{plot.hb_local}).
}
\description{
Simulates the local hydrological balance of a set of clusters of rice paddies
on a daily basis, under a set of ideal assumptions.

Prominently, paddies are assumed to be managed - \emph{i.e.} irrigated and drained
\itemize{
\item in such a way to stick as closely as possible to an established yearly
ideal plan for draining and irrigation (with corresponding ideal water
levels), while satisfying the constraint imposed by the fact that the total
daily outflow from all clusters is fixed (and known by independent
measurements).
}

More details on the concrete algorithm used in the simulations can be found
at TODO #46.

For analyzing Albufera data, you should not need to directly run this
function, and you can instead use the \link{albufera_hb_local} wrapper, that
calls \code{hb_local()} with the right arguments, extracted from the built-in
datasets.
}
\details{
All arguments of this function  should be conceptually thought as the columns
of a data-frame, each row of which corresponds to data for a given cluster
(specified by \code{cluster_id}) on a given day (specified by \code{date}). For this
reason, there should be a certain consistency among these vectors, which is
\strong{currently not being enforced}, and left to the correct specification from
the side of the user. To name some examples: \code{ditch} should be consistent
among \code{cluster_id}; \code{petp_cm} should be consistent among \code{date}; \emph{etc.etc.}.
}
