% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hba.R
\name{compute_hba}
\alias{compute_hba}
\alias{hba}
\title{Hydrological Balance of the Albufera Lake}
\usage{
compute_hba(
  model,
  storage_curve = linear_storage_curve(intercept = 16.7459 * 1e+06, slope = 23.6577 *
    1e+06),
  petp_surface = linear_petp_surface(surface_P = 114.225826072 * 1e+06, surface_ETP =
    79.360993685 * 1e+06)
)

hba(model, value = c("output", "params"))
}
\arguments{
\item{storage_curve}{a function that takes a numeric vector as input, and
returns a numeric vector of the same length. Function that converts lake
levels (passed through the \code{level} argument) into lake \emph{volumes}.}

\item{petp_surface}{a function that takes two numeric vectors of common
length as inputs, and returns a numeric vector of the same length. Function
that converts precipitation and evapotranspiration values (passed through
the \code{rain_mm} and \code{evapotranspiration_mm} arguments) into an overall volume change.}
}
\value{
An object of class \code{hba}, a lightweight wrapper of \code{data.frame}
with a few additional visualization methods (most prominently
\link{plot.hba}). The underlying data-frame contains as columns the
input time series, as well as the following calculated columns:
\itemize{
\item \code{volume} Volume time series, obtained from the storage curve.
\item \code{volume_change} Differenced time series of volume. The \eqn{n}-th is given
by \eqn{\Delta V _n \equiv V_{n+1}-V_n}, where \eqn{V_n} is volume at time
step \eqn{n}.
\item \code{petp_change} Time series. The amount \eqn{\Delta V _n ^\text{P-ETP}} of
volume change due to precipitation and evapotranspiration, obtained from the
P-ETP surface.
\item \code{outflow_total} Time series of total outflows, measured in cube meters per
second. This is the sum
\eqn{\sum _i O_i} of time series passed through the \code{outflows} argument,
plus an extra term \eqn{\delta O} of unaccounted outflow. The correction term
is defined by
\eqn{\delta O = \frac{\Delta V _n - \Delta V _n ^\text{P-ETP}}{24 \times 60 \times 60} -\sum _i O_i}
and is chosen in such a way to ensure that the total inflow is always
non-negative.
\item \code{outflow_extra} Time series. The unaccounted outflow term \eqn{\delta O}
described above.
\item \code{inflow_total} Time serie of total inflows, in cube meters per second.
This is computed as \eqn{I = \sum _{i} O_i + \delta O + \frac{\Delta V _n - \Delta V _n ^\text{P-ETP}}{24 \times 60 \times 60}}.
\item \code{residence_time_days}. Residence time, as modeled by \link{hba_residence_time}.
}
}
\description{
Wrapper around \link{hba}, used to run the global hydrological balance
calculations with the data for the Albufera lake packed up in the
\code{data.frame}s exported by \code{erahumed}.
}
\details{
The numeric inputs for the linear storage curve are taken from the CHJ report
\href{https://www.chj.es/Descargas/ProyectosOPH/Consulta\%20publica/PHC-2015-2021/ReferenciasBibliograficas/HumedalesZonasProtegidas/CHJ,2012.Aquatool_Albufera.pdf}{Modelo de seguimiento de l’Albufera de Valencia con AQUATOOLDMA.}.
The values used as the arguments of \code{petp_surface()} were calculated by the
package authors, and correspond to the total study area (\code{surface_P}) and
the flooded surface (\code{surface_ETP}).

The \code{outflows_df} data.frame is supposed to have all columns of
\link{albufera_outflows} whose names do not start by \code{outflow_}, with the
appropriate type. In addition, \code{outflows_df} can have an arbitrary number of
numeric columns named \verb{outflow_*}, which represent the measured outflows for
the system. It is fundamental that outflow columns follow this particualr
naming scheme, as these are automatically recognized by
\code{hba()} and passed down to low level functions.
}
\author{
Pablo Amador Crespo, Valerio Gherardi
}
