% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbd--main.R
\name{hbd}
\alias{hbd}
\alias{setup_hbd}
\title{Hydrological Balance of ditches}
\usage{
setup_hbd(simulation)
}
\value{
An object of class \link{erahumed_simulation}.
}
\description{
This layer simulates the hydrology of ditches of the Albufera lake.
}
\details{
No details
}
\seealso{
Other simulation layers: 
\code{\link{ca}},
\code{\link{ct}},
\code{\link{hba}},
\code{\link{hbp}},
\code{\link{inp}}
}
\concept{simulation layers}
