% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hb_global_units.R
\name{hbg_residence_time}
\alias{hbg_residence_time}
\title{Residence Time}
\usage{
hbg_residence_time(volume, outflow_total, k = 61, units = c("days", "seconds"))
}
\arguments{
\item{volume}{numeric vector. Time series of volumes in \eqn{\text{m}^3}.}

\item{outflow_total}{numeric vector. Time series of total outflow in
\eqn{\text{m}^3 / \text{s}}.}

\item{k}{positive integer. Size of the window in the moving average. The
default is of the order of magnitude of actual residence time for the
Albufera Lake.}

\item{units}{either \code{"days"} or \code{"seconds"}. Units of measure for
the returned time series.}
}
\value{
A numeric vector. Time series of residence times, in the units of
measure specified by the \code{units} argument (assuming \code{volume} and
\code{total_inflow} are provided in the correct units).
}
\description{
\emph{N.B.:}: This function is the one used internally by \link{hb_global} to
compute residence times, and is not directly exported by \code{{erahumed}}.

Computes residence times as
\eqn{t = s_k(\text{Volume}) / s_k(\vert \text{Outflow}\vert) },
where \eqn{s_k(\cdot)} denotes a moving average with a smoothing window of
size \eqn{k}, centered at the current observation, and
\eqn{\vert \cdot \vert} is the absolute value (to deal with cases in which
the inflow becomes negative).
}
\author{
Pablo Amador Crespo, Valerio Gherardi
}
