% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbl--logical-units.R
\name{hbl_residence_time}
\alias{hbl_residence_time}
\title{Residence Time}
\usage{
hbl_residence_time(volume, outflow_total, k = 61, units = c("days", "seconds"))
}
\arguments{
\item{volume}{numeric vector. Time series of volumes in \eqn{\text{m}^3}.}

\item{outflow_total}{numeric vector. Time series of total outflow in
\eqn{\text{m}^3 / \text{s}}.}

\item{k}{positive integer. Size of the window in the moving average. The
default is of the order of magnitude of actual residence time for the
Albufera Lake.}

\item{units}{either \code{"days"} or \code{"seconds"}. Units of measure for
the returned time series.}
}
\value{
A numeric vector. Time series of residence times, in the units of
measure specified by the \code{units} argument (assuming \code{volume} and
\code{total_inflow} are provided in the correct units).
}
\description{
\strong{N.B.:}: This function is the one used internally by \link{hbl} to
compute residence times, and is not directly exported by \code{{erahumed}}.

Computes residence times as
\eqn{t = \overline{\text{Volume}} / \overline{\text{Outflow}} },
where \eqn{\overline {\cdot}} denotes a (symmetric) moving average.
}
