% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbp--main.R
\name{hbp}
\alias{hbp}
\alias{compute_hbp}
\title{HBP: Hydrological Balance of rice Paddies}
\usage{
hbp(simulation)

compute_hbp(
  simulation,
  management_df = erahumed::albufera_management,
  clusters_df = erahumed::albufera_clusters,
  ideal_flow_rate_cm = 5,
  height_thresh_cm = 2
)
}
\arguments{
\item{simulation}{An object of class \link{erahumed_simulation}, with a pre-computed
\link{hba} layer (\emph{i.e.} such that \code{hba(simulation)} is not \code{NULL}).}

\item{management_df}{A \code{data.frame} that provides the yearly schedule for
irrigation and draining, that strictly follows the template of
\link{albufera_management}.}

\item{clusters_df}{A \code{data.frame} that contains the definition of rice paddy
clusters. This should strictly follow the template of
\link{albufera_clusters}, and it is in fact unlikely that the user would want
to change this input (TODO #47).}

\item{ideal_flow_rate_cm}{A positive number. Ideal inflow/outflow of a
cluster, for days in which the cluster is scheduled to be in flux (\emph{i.e.}
when being simultaneously irrigated and drained) Expressed in centimeters
per day.}

\item{height_thresh_cm}{A positive number. Height threshold for water levels,
below which a cluster is considered to be emptied.}
}
\value{
Objects of class \link{erahumed_simulation} and \code{erahumed_hbp}, for
\code{compute_hbp()} and \code{hbp()} respectively.
}
\description{
This simulation layer performs a simulation of hydrological balance
daily data for the rice paddies of the Albufera National Park, that provides
a synthetic dataset of daily inflows, outflows and average water levels for
each rice paddy.

The inputs to this simulation are:
\enumerate{
\item The previously calculated hydrological balance for the Albufera lake.
\item A geographical subdivision of the Albufera National Park into "clusters"
of rice paddies, of a definite rice variety, and whose waters flow into the
same ditch.
\item An yearly ideal schedule for irrigation and draining, with corresponding
expected water levels of a cluster for each day of the year.
}

This simulation layer requires the \link{hba} layer of the simulation to be
pre-computed.
}
\details{
TODO: #64, plus more detailed information on the structure of
\code{management_df} (and perhaps also on \code{clusters_df}, depending on the decision
taken regarding #47).
}
\seealso{
Other simulation layers: 
\code{\link{ca}()},
\code{\link{ct}()},
\code{\link{hba}()},
\code{\link{inp}()}
}
\author{
Pablo Amador Crespo, Valerio Gherardi
}
\concept{simulation layers}
