% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_balance.R
\name{hydro_balance_global}
\alias{hydro_balance_global}
\title{Global Hydrological Balance}
\usage{
hydro_balance_global(
  level,
  P,
  ETP,
  outflows,
  storage_curve = erahumed::linear_storage_curve,
  petp_curve = erahumed::petp_volume_change
)
}
\description{
Documentation TBD
}
