% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.R
\name{hydrology}
\alias{hydrology}
\alias{setup_hydrology}
\alias{compute_hydrology}
\title{Hydrology of the Albufera Natural Park}
\usage{
setup_hydrology(
  simulation,
  storage_curve = function(level) 25.58 * 1e+06 + level * 63.086 * 1e+06,
  petp_function = function(p, etp) 53.9 * 1000 * (p - etp),
  management_df = erahumed::albufera_management,
  ideal_flow_rate_cm = 5,
  height_thresh_cm = 0.5,
  ditch_level_m = 1
)

compute_hydrology(simulation)
}
\arguments{
\item{simulation}{\code{[}\link{erahumed_simulation}\verb{]} \cr
The simulation object being modified.}

\item{storage_curve}{\verb{[function]} \cr A function that converts lake levels into lake \emph{volumes}. The function should take a single numeric vector as input, and should return a numeric vector of the same length as output. The numeric values for the default (linear) storage curve are taken from the CHJ report \href{https://www.chj.es/Descargas/ProyectosOPH/Consulta\%20publica/PHC-2015-2021/ReferenciasBibliograficas/HumedalesZonasProtegidas/CHJ,2012.Aquatool_Albufera.pdf}{\emph{Modelo de seguimiento de l’Albufera de Valencia con AQUATOOLDMA}}.}

\item{petp_function}{\verb{[function]} \cr A function that converts values of precipitation and evapotranspiration per unit area (assumed to be expressed in millimiters) into volume changes for the Albufera lake (in cube meters when the inputs are given in the expected units). This function should take two numeric vectors of common length as inputs (representing the precipitation and evapotranspiration time series, respectively), and should return a numeric vector of the same length. The numeric values used in the default (linear) function were computed by the package authors, and correspond to the total study area (consisting of the lake and irrigation ditches) and the flooded fraction, that are relevant for precipitation and evapotranspiration volume changes, respectively.}

\item{management_df}{\verb{[data.frame]} \cr Dataset that provides the yearly schedule for irrigation and draining, along the template of \link{albufera_management} (the default value).}

\item{ideal_flow_rate_cm}{\verb{[numeric(1)]} \cr Ideal inflow/outflow of a cluster, for days in which the cluster is scheduled to be in flux (\emph{i.e.} when being simultaneously irrigated and drained). Expressed in centimeters per day.}

\item{height_thresh_cm}{\verb{[numeric(1)]} \cr A positive number. Height threshold for water levels, below which a cluster is considered to be emptied.}

\item{ditch_level_m}{\verb{[numeric(1)]} \cr Constant water depth in ditches.}
}
\value{
An \link{erahumed_simulation}.
}
\description{
These functions are used to setup the parameters and run the
calculations relevant for simulating the hydrology of the Albufera Natural
Park system.
}
