% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux--chemicals.R
\name{info_chemicals}
\alias{info_chemicals}
\title{Information on supported chemicals}
\usage{
info_chemicals()
}
\value{
A named list, each element of which corresponds to a chemical.
}
\description{
This helper provides information on the set of chemicals supported by
ERAHUMED.
}
