% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux--clusters.R
\name{info_clusters}
\alias{info_clusters}
\title{Information on rice field clusters}
\source{
Martínez-Megías, C., Mentzel, S., Fuentes-Edfuf, Y., Moe, S. J., &
Rico, A. (2023). Influence of climate change and pesticide use practices on
the ecological risks of pesticides in a protected Mediterranean wetland: A
Bayesian network approach. \emph{Science of The Total Environment, 878},
163018. \url{https://doi.org/10.1016/j.scitotenv.2023.163018}
}
\usage{
info_clusters(include_geometry = FALSE)
}
\arguments{
\item{include_geometry}{\code{TRUE} or \code{FALSE}. Whether to include the geometries
of clusters (as a column of class \code{sfc_MULTIPOLYGON} from the \code{{sf}}
package).}
}
\value{
A \code{data.frame}.
}
\description{
This helper returns a \code{data.frame} with the list of rice field clusters
of the Albufera Natural Park used by ERAHUMED. The definition of
the clusters is discussed
\href{https://doi.org/10.1016/j.scitotenv.2023.163018}{Martínez-Megías et al. (2023)}.
}
