% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inp.R
\name{compute_inp}
\alias{compute_inp}
\alias{inp}
\title{ERAHUMED Input Data}
\usage{
compute_inp(
  model,
  outflows_df = erahumed::albufera_outflows,
  petp_df = erahumed::albufera_petp
)

inp(model, value = c("output", "params"))
}
\arguments{
\item{outflows_df, petp_df}{\code{data.frame}s, whose structures follow the
templates of \link{albufera_outflows}, \link{albufera_petp}.}
}
\value{
An object of class \code{inp}.
}
\description{
Input data for the algorithms of the ERAHUMED DSS.
}
\details{
TODO
}
