% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dss--api.R
\name{launch_dss}
\alias{launch_dss}
\title{ERAHUMED Decision Support System Graphical User Interface}
\usage{
launch_dss()
}
\description{
Opens the graphical user interface (GUI) of the ERAHUMED Decision Support
System. This is a Shiny application that provides access to the package's
functionalities through a point-and-click interface, and can be used as an
alternative to calling the functions programmatically.
}
