% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta--layer-parameters.R
\name{layer_parameters}
\alias{layer_parameters}
\alias{layer_parameters.erahumed_simulation_layer}
\alias{layer_parameters.erahumed_simulation}
\title{Extract simulation layer parameters}
\usage{
layer_parameters(object, ...)

\method{layer_parameters}{erahumed_simulation_layer}(object, ...)

\method{layer_parameters}{erahumed_simulation}(object, layer, ...)
}
\arguments{
\item{object}{The object from which the layer parameters are to be
extracted from (see above for available methods).}

\item{...}{further arguments passed to or from other methods.}

\item{layer}{A string. Name of the layer whose parameters \code{list}
is to be extracted.}
}
\value{
A \code{list}.
}
\description{
Extracts the parameters used in the computation of a simulation layer.
}
\author{
Valerio Gherardi
}
