% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hba--logical-units.R
\name{linear_petp_surface}
\alias{linear_petp_surface}
\title{Volume daily changes due to Precipitation and Evapotranspiration}
\usage{
linear_petp_surface(surface_P, surface_ETP)
}
\arguments{
\item{surface_P, surface_ETP}{positive numbers. The effective surfaces (
expressed in \eqn{\text{m}^2}) for volume changes due to precipitation and
evapotranspiration (see description).}
}
\value{
A function with signature \verb{function(P, ETP)}, whose arguments are
assumed to be numeric vectors, that computes volumes changes due to
precipitation and evapotranspiration, according to the prescribed linear
relationship.
}
\description{
Creates a (homogeneous) linear fucntion that computes volume daily changes
due to precipitation and evapotranspiration
as \eqn{\Delta V = \Delta h_\text{P} \cdot S_\text{P} - \Delta h_\text{ETP} \cdot S_\text{ETP}}, where
\eqn{\Delta h _\text{P,ETP}} are precipitation and evapotranspiration in
expressed in millimiters, and \eqn{S_{\text{P,ETP}}} are effective surfaces.
}
\author{
Pablo Amador Crespo, Valerio Gherardi
}
