% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhb_units.R
\name{linear_storage_curve}
\alias{linear_storage_curve}
\title{Linear Storage Curve}
\usage{
linear_storage_curve(intercept, slope)
}
\arguments{
\item{intercept}{a number. Intercept of the linear storage curve.}

\item{slope}{a number. Slope of the linear storage curve}
}
\value{
A function with signature \verb{function(level)}, whose argument is
assumed to be a numeric vector, that computes volumes from lake levels
according to the prescribed linear relationship.
}
\description{
Creates a linear storage curve with a given slope and intercept,
\emph{viz.} \eqn{\text{Volume} = \text{slope} \times \text{Level} + \text{intercept}}.
}
\details{
The handling of units of measurement is left to the user.
}
\author{
Pablo Amador Crespo, Valerio Gherardi
}
