% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meteo_beni_2023}
\alias{meteo_beni_2023}
\title{Meteo Beni 2023}
\format{
\subsection{\code{meteo_beni_2023}}{

A data frame with 6140 rows and 5 columns:
\describe{
\item{A}{??}
\item{B}{??}
\item{C}{??}
}
}
}
\source{
???
}
\usage{
meteo_beni_2023
}
\description{
???
}
\keyword{datasets}
