% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paddy_management}
\alias{paddy_management}
\title{Paddy Management Data}
\format{
\subsection{\code{paddy_management}}{

A data frame with R rows and C columns:
\describe{
\item{A}{??}
\item{B}{??}
\item{C}{??}
}
}
}
\source{
???
}
\usage{
paddy_management
}
\description{
TBD.
}
\keyword{datasets}
