% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem--presets.R
\name{pesticides}
\alias{pesticides}
\alias{acetamiprid}
\alias{azoxystrobin}
\alias{bentazone}
\alias{cycloxydim}
\alias{cyhalofop_butyl}
\alias{difenoconazole}
\alias{mcpa}
\alias{penoxsulam}
\title{Predefined pesticide definitions}
\usage{
acetamiprid()

azoxystrobin()

bentazone()

cycloxydim()

cyhalofop_butyl()

difenoconazole()

mcpa()

penoxsulam()
}
\value{
An object of class \link{chemical}.
}
\description{
These functions return predefined objects of class \link{chemical},
representing commonly used pesticides and their properties.
}
\examples{
acetamiprid()

}
