% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidro_balance.R
\name{petp_volume_change}
\alias{petp_volume_change}
\title{Volume daily changes due to Precipitation and Evapotranspiration}
\usage{
petp_volume_change(
  P,
  ETP,
  surface_P = 114.225826072 * 1e+06,
  surface_ETP = 79.360993685 * 1e+06
)
}
\arguments{
\item{P, ETP}{numeric vectors of same length. Values of precipitation and
evapotranspiration in \eqn{\text{mm}}.}

\item{surface_P, surface_ETP}{positive numbers. The effective surfaces (
expressed in \eqn{\text{m}^2}) for volume changes due to precipitation and
evapotranspiration (see description).}
}
\value{
A dataframe containing the same columns as \code{input}, and an additional
column (see \code{output_col} argument) with the computed values of volume
daily changes, expressed in \eqn{\text{m} ^3 / \text{s}}.
}
\description{
Computes the volume daily changes due to precipitation and evapotranspiration
as \eqn{\Delta V = \Delta h_\text{P} \cdot S_\text{P} - \Delta h_\text{ETP} \cdot S_\text{ETP}}, where
\eqn{\Delta h _\text{P,ETP}} are precipitation and evapotranspiration in
expressed in millimiters, and \eqn{S_{\text{P,ETP}}} are effective surfaces.
}
\details{
TODO: document the extraction of effective surface values.
TODO: other options to obtain volume changes?
}
\author{
Pablo Amador Crespo, Valerio Gherardi
}
