% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca--plot.R
\name{plot.erahumed_ca}
\alias{plot.erahumed_ca}
\title{Plot CA simulation layer output}
\usage{
\method{plot}{erahumed_ca}(x, type = c("cluster_view", "timeline_view"), ...)
}
\arguments{
\item{x}{An object of class \code{ca}.}

\item{type}{Type of plot to be generated. Currently, only the
\code{"cluster_view"} plot type is implemented.}

\item{...}{Further plotting parameters, associated with the plot type
specified by \code{type}; See details.}
}
\value{
A plotly plot.
}
\description{
Plot method for \link{ca} simulation layers.
}
\details{
The \code{"cluster_view"} plot generates a time series plot of the daily
water levels (in cm) of an individual cluster, with superimposed lines
indicating applications. In order to use this plotting
method, the user must provide an additional \code{cluster_id} argument, a string
specifying the identifier of the cluster whose levels are to be plotted.
}
