% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct--plot.R
\name{plot.erahumed_ct}
\alias{plot.erahumed_ct}
\title{Plot CT model component output}
\usage{
\method{plot}{erahumed_ct}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ct}.}

\item{...}{Further plotting parameters.}
}
\value{
A plotly plot.
}
\description{
Plot method for \link{ct} model components.
}
\details{
TBD.
}
