% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctc--plot.R
\name{plot.erahumed_ctc}
\alias{plot.erahumed_ctc}
\title{Plot ctc simulation layer output}
\usage{
\method{plot}{erahumed_ctc}(
  x,
  type = c("cluster_view", "max_boxplot"),
  variable = c("mass", "density"),
  ...
)
}
\arguments{
\item{x}{An object of class \code{ctc}.}

\item{type}{Type of plot to be generated. Currently, only the
\code{"cluster_view"} plot type is implemented.}

\item{variable}{Variable to plot for \code{"cluster_view"} type plots, either
\code{"mass"} or \code{"density"}; ignored for other types of plots.}

\item{...}{Further plotting parameters.}
}
\value{
A \link[dygraphs]{dygraph} plot.
}
\description{
Plot method for \link{ctc} simulation layers.
}
\details{
The \code{"cluster_view"} plot produces a time series plot of the daily
total amounts or densities of chemical (in Kg) for an individual cluster.
In order to use this plotting method, the user must provide an additional
\code{cluster_id} argument, a string specifying the identifier of the cluster
whose levels are to be plotted.

The \code{"max_boxplot"} type plot produces a box plot of the maximum chemical
densities (in Kg per cube meter) present in sediment, water and outflow.
}
