% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hba--plot.R
\name{plot.erahumed_hba}
\alias{plot.erahumed_hba}
\title{Plot HBA model layer output}
\usage{
\method{plot}{erahumed_hba}(x, variable, ...)
}
\arguments{
\item{x}{An object of class \code{hba}.}

\item{variable}{The variable to be plotted. Can be any numeric column of
\code{layer_output(x)}.}

\item{...}{Not used.}
}
\value{
A plotly plot.
}
\description{
Plot method for \link{hba} model layers.
}
