% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbp--plot.R
\name{plot.erahumed_hbc}
\alias{plot.erahumed_hbc}
\title{Plot hbc simulation layer output}
\usage{
\method{plot}{erahumed_hbc}(x, type = c("cluster_view", "map_view"), ...)
}
\arguments{
\item{x}{An object of class \code{hbc}.}

\item{type}{Type of plot to be generated. Currently, only the
\code{"cluster_view"} plot type is implemented.}

\item{...}{Further plotting parameters, associated with the plot type
specified by \code{type}; See details.}
}
\value{
A \link[dygraphs]{dygraph} plot.
}
\description{
Plot method for \link{hbc} simulation layers.
}
\details{
The \code{"cluster_view"} plot generates a time series plot of the daily
water levels (in cm) of an individual cluster. In order to use this plotting
method, the user must provide an additional \code{cluster_id} argument, a string
specifying the identifier of the cluster whose levels are to be plotted.
}
