% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbd--plot.R
\name{plot.erahumed_hbd}
\alias{plot.erahumed_hbd}
\title{Plot HBD simulation layer output}
\usage{
\method{plot}{erahumed_hbd}(x, ditch = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{hbd}.}

\item{ditch}{string. Ditch to be plotted, one of the values of the \code{ditch }
column in the \code{info_ditches()} output.}

\item{...}{Not used}
}
\value{
A \link[dygraphs]{dygraph} plot.
}
\description{
Plot method for \link{hbd} simulation layers.
}
