% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbl--plot.R
\name{plot.erahumed_hbl}
\alias{plot.erahumed_hbl}
\title{Plot hbl simulation layer output}
\usage{
\method{plot}{erahumed_hbl}(x, variable = "inflow_total", ...)
}
\arguments{
\item{x}{An object of class \code{hbl}.}

\item{variable}{The variable to be plotted. Can be any numeric column of
\code{get_layer_output(x)}.}

\item{...}{Not used.}
}
\value{
A \link[dygraphs]{dygraph} plot.
}
\description{
Plot method for \link{hbl} simulation layers.
}
