% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbp_plot.R
\name{plot.erahumed_hbp}
\alias{plot.erahumed_hbp}
\title{Plotting local hydrological balance data}
\usage{
\method{plot}{erahumed_hbp}(x, type = c("cluster_levels", "map"), ...)
}
\arguments{
\item{x}{The object of class \code{hba} containing the data to be plotted.}

\item{type}{Type of plot to be generated. Currently, only the
\code{"cluster_levels"} plot is implemented.}

\item{...}{Further plotting parameters, associated with the plot type
specified by \code{type}; See details.}
}
\value{
A plotly plot.
}
\description{
Plot method for local Hydrological balance data, generated through
\link{hbp}. A simple wrapper around \link[plotly]{plot_ly} to generate
time series plots of the calculated quantities.
}
\details{
The \code{"cluster_levels"} plot generates a time series plot of the daily
water levels (in cm) of an individual cluster. In order to use this plotting
method, the user must provide an additional \code{cluster_id} argument, a string
specifying the identifier of the cluster whose levels are to be plotted (see
example below).
}
