% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inp--plot.R
\name{plot.erahumed_inp}
\alias{plot.erahumed_inp}
\title{Plot INP model component output}
\usage{
\method{plot}{erahumed_inp}(x, ...)
}
\arguments{
\item{x}{An object of class \code{inp}.}

\item{...}{Further plotting parameters; see details.}
}
\value{
A plotly plot.
}
\description{
\strong{Not yet implemented.} Plot method for \link{inp} model components.
}
\details{
\strong{Not yet implemented.}
}
