% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hb_plot.R
\name{plot.hb_global}
\alias{plot.hb_global}
\title{Plotting global hydrological balance data}
\usage{
\method{plot}{hb_global}(x, variable, ...)
}
\arguments{
\item{x}{The object of class \code{hb_global} containing the data to be plotted.}

\item{variable}{The variable to be plotted. Can be any numeric column of \code{x}.}

\item{...}{Not used.}
}
\value{
A plotly plot.
}
\description{
Plot method for global hydrological balance data, generated through
\link{hb_global}. A simple wrapper around \link[plotly]{plot_ly} to generate
time series plots of the calculated quantities.
}
