% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hb_plot.R
\name{plot.hb_local}
\alias{plot.hb_local}
\title{Plotting local hydrological balance data}
\usage{
\method{plot}{hb_local}(x, type = c("cluster_levels", "map"), ...)
}
\arguments{
\item{x}{The object of class \code{hb_global} containing the data to be plotted.}

\item{type}{Type of plot to be generated. Currently, only the
\code{"cluster_levels"} plot is implemented.}

\item{...}{Further plotting parameters, associated with the plot type
specified by \code{type}; See details.}
}
\value{
A plotly plot.
}
\description{
Plot method for local Hydrological balance data, generated through
\link{hb_local}. A simple wrapper around \link[plotly]{plot_ly} to generate
time series plots of the calculated quantities.
}
\details{
The \code{"cluster_levels"} plot generates a time series plot of the daily
water levels (in cm) of an individual cluster. In order to use this plotting
method, the user must provide an additional \code{cluster_id} argument, a string
specifying the identifier of the cluster whose levels are to be plotted (see
example below).
}
\examples{
plot(albufera_hb_local(),
     type = "cluster_levels",
     cluster_id = "02_Carrera_del_Saler0-2_0"
     )

}
