% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hb_utils.R
\name{residence_time}
\alias{residence_time}
\title{Residence Time}
\usage{
residence_time(volume, inflow_total, k = 61, units = c("days", "seconds"))
}
\arguments{
\item{volume}{numeric vector. Time series of volumes in \eqn{\text{m}^3}.}

\item{k}{positive integer. Size of the window in the moving average. The
default is of the order of magnitude of actual residence time for the
Albufera Lake.}

\item{units}{either \code{"days"} or \code{"seconds"}. Units of measure for
the returned time series.}

\item{total_inflow}{numeric vector. Time series of total inflow in
\eqn{\text{m}^3 / \text{s}}.}
}
\value{
A numeric vector. Time series of residence times, in the units of
measure specified by the \code{units} argument (assuming \code{volume} and
\code{total_inflow} are provided in the correct units).
}
\description{
Computes residence times as
\eqn{t = s_k(\text{Volume}) / s_k(\vert \text{Inflow}\vert) },
where \eqn{s_k(\cdot)} denotes a moving average with a smoothing window of
size \eqn{k}, centered at the current observation, and
\eqn{\vert \cdot \vert} is the absolute value (to deal with cases in which
the inflow becomes negative).
}
\author{
Pablo Amador Crespo, Valerio Gherardi
}
