% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfms--docs.R, R/rfms--presets.R,
%   R/rfms--s3-class.R
\name{rfms}
\alias{rfms}
\alias{jsendra}
\alias{bomba}
\alias{clearfield}
\alias{new_rfms}
\title{Rice Field Management System}
\usage{
jsendra()

bomba()

clearfield()

new_rfms(
  sowing_yday = 113,
  harvesting_yday = 251,
  perellona_end_yday = 15,
  perellona_start_yday = 306,
  flow_height_cm = 10,
  perellona_height_cm = 20,
  display_name = "New Management System"
)
}
\arguments{
\item{sowing_yday}{\verb{[integer(1)]} \cr Day of the year marking the start of the sowing season (1–366, assuming a leap year).}

\item{harvesting_yday}{\verb{[integer(1)]} \cr Day of the year marking the end of the sowing season (1–366, assuming a leap year).}

\item{perellona_end_yday}{\verb{[integer(1)]} \cr Day of the year marking the end of the \emph{Perellona} flooding period (before sowing).}

\item{perellona_start_yday}{\verb{[integer(1)]} \cr Day of the year marking the beginning of the \emph{Perellona} flooding period (after harvest).}

\item{flow_height_cm}{\verb{[numeric(1)]} \cr Target water level (in cm) during the regular days of the sowing season, excluding emptying and transition days.}

\item{perellona_height_cm}{\verb{[numeric(1)]} \cr Target water level (in cm) during the \emph{Perellona} flooding period.}

\item{display_name}{\verb{[character(1)]} \cr Name of the management system to be displayed in output plots and summaries.}
}
\value{
An object of class \code{erahumed_rfms}.
}
\description{
Functions to define and initialize rice field management systems (RFMSs).
}
\details{
These functions are used to define and initialize rice field
management systems. Specifically:
\itemize{
\item \code{new_rfms()} creates an empty management system, with no
scheduled chemical applications.
\item \code{jsendra()}, \code{bomba()}, and \code{clearfield()} provide
predefined systems inspired by the \emph{J. Sendra}, \emph{Bomba}, and \emph{Clearfield}
rice varieties, respectively.
}

Additional chemical applications can be scheduled using the helper function
\code{\link[=schedule_application]{schedule_application()}}. The default values of the arguments of
\code{new_rfms()} coincide with those used internally by \code{jsendra()},
\code{bomba()}, and \code{clearfield()}.

For a detailed explanation of RFMS concepts, configuration, and
built-in presets, see the
\href{https://erahumed.github.io/erahumed-book/chapters/rfms.html}{RFMS section of the user manual}.
}
\seealso{
\link{schedule_application}
}
