% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk}
\alias{risk}
\alias{setup_risk}
\alias{compute_risk}
\title{Ecological risk in the Albufera Natural Park}
\usage{
setup_risk(simulation)

compute_risk(simulation)
}
\arguments{
\item{simulation}{\code{[}\link{erahumed_simulation}\verb{]} \cr
The simulation object being modified.}
}
\value{
Functions \verb{setup_*()} and \verb{compute_*()} return a
\link{erahumed_simulation}, while \verb{extract_*()} returns a \code{data.frame}.
}
\description{
These functions are used to setup, run and extract simulation
results for ecological risk in the Albufera Natural Park system.
}
