% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfms--s3-class.R
\name{schedule_application}
\alias{schedule_application}
\title{Schedule chemical application}
\usage{
schedule_application(
  system,
  chemical,
  amount_kg_ha,
  seed_day,
  type = c("ground", "aerial"),
  emptying_days = 1
)
}
\arguments{
\item{system}{\verb{[erahumed_rfms]} \cr A rice field management system to which the application is added. See \link{rfms} for details.}

\item{chemical}{\verb{[erahumed_chemical]} \cr Chemical compound to be applied. See \link{chemical} for details.}

\item{amount_kg_ha}{\verb{[numeric(1)]} \cr Amount of active ingredient applied per hectare.}

\item{seed_day}{\verb{[integer(1)]} \cr Number of days after seeding when the application is performed.}

\item{type}{\verb{[character(1)]} \cr Application method, either 'ground' (requiring draining) or 'aerial'.}

\item{emptying_days}{\verb{[numeric(1)]} \cr Number of days the field remains empty \strong{before} a ground application. The chemical is always assumed to be applied on the \strong{last day} of this emptying period. Ignored if the application type is \code{"aerial"}.}
}
\value{
An updated object of class \link[=rfms]{erahumed_rfms} with the scheduled application included.
}
\description{
Adds a chemical application to a rice field management system (RFMS), specifying timing, application method, and chemical properties.
}
\seealso{
\link{rfms}, \link{chemical}
}
