% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hb_global_wrapper.R
\name{albufera_hb_global}
\alias{albufera_hb_global}
\title{Albufera Global Hydrological Balance}
\usage{
albufera_hb_global(
  outflows_df = erahumed::albufera_outflows,
  petp_df = erahumed::albufera_petp,
  storage_curve = linear_storage_curve(intercept = 16.7459 * 1e+06, slope = 23.6577 *
    1e+06),
  petp_surface = linear_petp_surface(surface_P = 114.225826072 * 1e+06, surface_ETP =
    79.360993685 * 1e+06)
)
}
\arguments{
\item{outflows_df, petp_df}{\code{data.frame}s, whose structures follow the
templates of \link{albufera_outflows} and \link{albufera_petp}, respectively.}

\item{storage_curve}{a function that takes a numeric vector as input, and
returns a numeric vector of the same length. Function that converts lake
levels (passed through the \code{level} argument) into lake \emph{volumes}.}

\item{petp_surface}{a function that takes two numeric vectors of common
length as inputs, and returns a numeric vector of the same length. Function
that converts precipitation and evapotranspiration values (passed through
the \code{rain_mm} and \code{evapotranspiration_mm} arguments) into an overall volume change.}
}
\value{
Same as \link{hb_global}.
}
\description{
Wrapper around \link{hb_global}, used to run the global hydrological balance
calculations with the data for the Albufera lake packed up in the
\code{data.frame}s exported by \code{erahumed}.
}
\details{
The numeric inputs used from the linear storage curve and P-ETP surface were
extracted from the CHJ report
\href{https://www.chj.es/Descargas/ProyectosOPH/Consulta\%20publica/PHC-2015-2021/ReferenciasBibliograficas/HumedalesZonasProtegidas/CHJ,2012.Aquatool_Albufera.pdf}{Modelo de seguimiento de l’Albufera de Valencia con AQUATOOLDMA.}
}
