% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{albufera_management}
\alias{albufera_management}
\title{Albufera Management Data}
\format{
\subsection{\code{albufera_management}}{

The dataset has one row per day of year for all combinations of the
categorical variables \code{tancat} and \code{variety}. The columns are the following:
\describe{
\item{mm}{numeric. Month of year (1 = January, 2 = February, \emph{etc.}).}
\item{dd}{numeric. Day of month.}
\item{tancat}{logical. Whether the paddy is a "tancat" or not.}
\item{variety}{character. Variety of rice planted in the paddy under
consideration.}
\item{sowing}{logical. Whether \code{mm} and \code{dd} correspond to the sowing day.}
\item{irrigation}{logical. Whether the paddy is scheduled to be irrigated
on this day.}
\item{draining}{logical. Whether the paddy is scheduled to be drained on
this day.}
\item{height_cm}{numeric. Water level of the paddy.}
}
}
}
\source{
\url{https://doi.org/10.1016/j.scitotenv.2023.163018}
}
\usage{
albufera_management
}
\description{
Data on management schedules of rice paddies in the Albufera National Park.
Contains information on irrigation, draining and water level of paddies by
day of year.
}
\keyword{datasets}
