% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs--data.R
\docType{data}
\name{albufera_management}
\alias{albufera_management}
\title{Albufera Management Data}
\format{
\subsection{\code{albufera_management}}{

The cardinality of data is given by: one row per day of year (29th of Feb. included) and per combination of the categorical variables \code{tancat} and \code{variety}.The dataset features the following columns:\describe{\item{mm}{numeric. Month of year (1 = January, 2 = February, \emph{etc.}).}\item{dd}{numeric. Day of month.}\item{tancat}{logical. Whether the paddy is a tancat or not.}\item{variety}{character. Rice variety of the paddy under consideration.}\item{sowing}{logical. Whether \code{mm} and \code{dd} correspond to the sowing day.}\item{ideal_irrigation}{logical. Whether the paddy is scheduled to be irrigated on this day.}\item{ideal_draining}{logical. Whether the paddy is scheduled to be drained on this day.}\item{ideal_height_eod_cm}{numeric. Scheduled water level of the paddy \emph{at the end of the day} (that is, after irrigation and draining).}}
}
}
\source{
Martínez-Megías, C., Mentzel, S., Fuentes-Edfuf, Y., Moe, S. J., &
Rico, A. (2023). Influence of climate change and pesticide use practices on
the ecological risks of pesticides in a protected Mediterranean wetland: A
Bayesian network approach. \emph{Science of The Total Environment, 878},
163018. \url{https://doi.org/10.1016/j.scitotenv.2023.163018}
}
\usage{
albufera_management
}
\description{
Data on management schedules of rice paddies in the Albufera National Park.
Contains information on irrigation, draining and water level of paddies by
day of year. For further details on how this information was obtained, see
the original source
\href{https://doi.org/10.1016/j.scitotenv.2023.163018}{Martínez-Megías et al. (2023)}.
}
\keyword{datasets}
