% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{albufera_outflows}
\alias{albufera_outflows}
\title{Albufera Outflows Daily Data}
\format{
\subsection{\code{albufera_outflows}}{

A data frame with 6140 rows and 5 columns:
\describe{
\item{Date}{Date of measurement}
\item{Level}{Lake level (in meters above sea level)}
\item{Pujol}{Outflow at Pujol (meters cube per second)}
\item{Perellonet}{Outflow at Perellonet (meters cube per second)}
\item{Perello}{Outflow at Perello (meters cube per second)}
}
}
}
\source{
\url{https://aps.chj.es/}
}
\usage{
albufera_outflows
}
\description{
Data from continuous measurements of outflow rates and lake level from
"Confederación Hidrográfica del Júcar".
}
\details{
Datos de los medidores en continuo de las salidas del lago.
}
\keyword{datasets}
