% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{albufera_outflows}
\alias{albufera_outflows}
\title{Albufera Outflows Daily Data}
\format{
\subsection{\code{albufera_outflows}}{

The dataframes have the following columns:
\describe{
\item{date}{Date of measurement}
\item{level}{Lake level (in meters above sea level)}
\item{pujol}{Outflow at Pujol (meters cube per second)}
\item{perellonet}{Outflow at Perellonet (meters cube per second)}
\item{perello}{Outflow at Perello (meters cube per second)}
\item{level_is_imputed}{Whether the \code{level} value was imputed.}
\item{pujol_is_imputed}{Whether the \code{pujol} value was imputed.}
\item{perellonet_is_imputed}{Whether the \code{perellonet} value was imputed.}
\item{perello_is_imputed}{Whether the \code{perello} value was imputed.}
}
}
}
\source{
\url{https://aps.chj.es/}
}
\usage{
albufera_outflows
}
\description{
Data from continuous measurements of outflow rates and lake level from
"Confederación Hidrográfica del Júcar" (CHJ). The original data has several
missing entries, which have been imputed in this dataset using an approach
based on GAMs.

TODO:
\enumerate{
\item Improve the documentation of imputation approach;
}
}
\keyword{datasets}
