% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs--data.R
\docType{data}
\name{albufera_outflows}
\alias{albufera_outflows}
\title{Albufera Outflows Daily Data}
\format{
\subsection{\code{albufera_outflows}}{

The cardinality of data is given by: one row per day in the desired study frame.The dataset features the following columns:\describe{\item{date}{Date of measurement}\item{level}{Lake level (meters above sea level)}\item{outflow_pujol}{Outflow at Pujol (cubic meters per second)}\item{outflow_perellonet}{Outflow at Perellonet (cubic meters per second)}\item{outflow_perello}{Outflow at Perello (cubic meters per second)}}
}
}
\source{
\url{https://aps.chj.es/}
}
\usage{
albufera_outflows
}
\description{
Data from continuous measurements of outflow rates and lake levels from
"Confederación Hidrográfica del Júcar" (CHJ).
}
\seealso{
Other input data: 
\code{\link{albufera_weather}}
}
\concept{input data}
\keyword{datasets}
