% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{albufera_petp}
\alias{albufera_petp}
\title{Albufera Precipitation and Evapotranspiration Daily Data}
\format{
\subsection{\code{albufera_petp}}{

A data frame with 8552 rows and 3 columns:
\describe{
\item{date}{Date of measurement}
\item{rain_mm}{Daily precipitation in millimiters.}
\item{evapotranspiration_mm}{Daily evapotranspiration in millimiters.}
}
}
}
\source{
TODO
}
\usage{
albufera_petp
}
\description{
TODO
}
\keyword{datasets}
