% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{albufera_weather}
\alias{albufera_weather}
\title{Albufera Weather}
\format{
\subsection{\code{albufera_weather}}{

A data frame with 6140 rows and 5 columns:
\describe{
\item{A}{??}
\item{B}{??}
\item{C}{??}
}
}
}
\source{
???
}
\usage{
albufera_weather
}
\description{
Albufera Weather
}
\keyword{datasets}
