% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs--data.R
\docType{data}
\name{albufera_weather}
\alias{albufera_weather}
\title{Albufera Weather Daily Data}
\format{
\subsection{\code{albufera_weather}}{

The cardinality of data is given by: one row per day in the desired study frame.The dataset features the following columns:\describe{\item{date}{Date of measurement}\item{temperature_ave_celsius}{Daily average temperature (degrees Celsius)}\item{temperature_min_celsius}{Daily minimum temperature (degrees Celsius)}\item{temperature_max_celsius}{Daily maximum temperature (degrees Celsius).}\item{precipitation_mm}{Daily precipitation (millimeters).}\item{evapotranspiration_mm}{Daily evapotranspiration (millimeters).}}
}
}
\source{
\url{http://riegos.ivia.es/}
}
\usage{
albufera_weather
}
\description{
Measurements from the Benifaió meteorological station operated
by the \emph{Instituto Valenciano de Investigaciones Agrarias} (IVIA),
considered as representative data for the Albufera Natural Park.
Publicly available at \url{http://riegos.ivia.es/}.
}
\seealso{
Other input data: 
\code{\link{albufera_outflows}}
}
\concept{input data}
\keyword{datasets}
