% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca--main.R
\name{ca}
\alias{ca}
\alias{compute_ca}
\title{CA: Chemical Applications}
\usage{
ca(simulation)

compute_ca(simulation, ca_schedules_df = erahumed::albufera_ca_schedules)
}
\arguments{
\item{simulation}{An object of class \link{erahumed_simulation}, with a pre-computed
\link{hbp} layer (\emph{i.e.} such that \code{hbp(simulation)} is not \code{NULL}).}

\item{ca_schedules_df}{a \code{data.frame} following the template of
\link{albufera_ca_schedules}. Each row of this data.frame corresponds to a
scheduled application. The semantics of columns are the same as in
\link{albufera_ca_schedules}.}
}
\value{
Objects of class \link{erahumed_simulation} and \code{erahumed_ca}, for
\code{compute_ca()} and \code{ca()} respectively.
}
\description{
This simulation layer simulates the application of chemicals to rice paddy
clusters, based on a previously computed simulation of local hydrological
balance. The result is a set of time series of applied doses, one for each
applied chemical.

This simulation layer requires the \link{hbp} layer of the simulation to be
pre-computed.
}
\details{
The output \code{data.frame} extends the output of the underlying \link{hbp}
layer, preserving its cardinality (one row per cluster per day). The
additional columns, named as the chemicals appearing in \code{ca_schedules_df},
provide the time series of applied doses, expressed in kilograms.
}
\seealso{
Other simulation layers: 
\code{\link{ct}()},
\code{\link{hba}()},
\code{\link{hbp}()},
\code{\link{inp}()}
}
\author{
Pablo Amador Crespo, Valerio Gherardi
}
\concept{simulation layers}
