% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca.R
\name{ca}
\alias{ca}
\title{Chemicals Applications}
\usage{
ca(model)
}
\arguments{
\item{hbp_res}{an object of class \code{"hbp"} - see \link{hbp}.}

\item{ca_schedules_df}{a \code{data.frame} following the template of
\link{albufera_ca_schedules}.}

\item{height_thresh_cm}{a positive number. Upper limit of paddy water levels
required for ground applications of chemicals. Expressed in centimeters.}
}
\value{
an object of class \code{"erahumed_ca"}. This is essentially a
\code{data.frame} with the same columns of the output of \link{hbp}(),
plus additional columns named as the chemicals listed in \code{ca_schedules_df},
each of which provides the (daily) time series of chemicals applications,
expressed in kilograms.
}
\description{
Simulates the application of chemicals to rice paddies, given a previous
calculation of local hydrological balance as input.
}
