% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca--main.R
\name{ca}
\alias{ca}
\alias{setup_ca}
\title{Chemical Applications}
\usage{
setup_ca(simulation, ca_schedules_df)
}
\arguments{
\item{simulation}{\code{[}\link{erahumed_simulation}\verb{]} \cr
The simulation object being modified.}

\item{ca_schedules_df}{\verb{[data.frame]} \cr A dataset that provides the list of scheduled chemical applications, along the template of \link{albufera_ca_schedules}.}
}
\value{
An objects of class \link{erahumed_simulation}.
}
\description{
This layer computes the application days of chemicals to rice fields, in agreement with the simulated hydrological balance of the latters.
}
\details{
The output \code{data.frame} extends the output of the underlying \link{hbp}
layer, preserving its cardinality (one row per cluster per day). The
additional columns, named as the chemicals appearing in \code{ca_schedules_df},
provide the time series of applied doses, expressed in kilograms.
}
\seealso{
Other simulation layers: 
\code{\link{ct}}
}
\concept{simulation layers}
