% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca.R
\name{ca}
\alias{ca}
\title{Chemicals Applications}
\usage{
ca(
  hbl,
  ca_schedules_df = erahumed::albufera_ca_schedules,
  height_thresh_cm = 2.5,
  sowing_mmdd = "04-20"
)
}
\arguments{
\item{hbl}{an object of class \code{"hb_local"} - see \link{hb_local}.}

\item{ca_schedules_df}{a \code{data.frame} following the template of
\link{albufera_ca_schedules}.}

\item{height_thresh_cm}{a positive number. Upper limit of paddy water levels
required for ground applications of chemicals. Expressed in centimeters.}

\item{sowing_mmdd}{a string formatted as \code{"\%m-\%d"}, where \verb{\%m} and \verb{\%d}
represent month and day, respectively. Day of the year at which sowing
occurs.}
}
\value{
an object of class \code{"erahumed_ca"}. This is essentially a
\code{data.frame} similar to the output of \link{hb_local}, with additional
columns named as the chemicals listed in \code{ca_schedules_df}. Each of these
additional columns provides the (daily) time series of chemicals
applications.
}
\description{
Simulates the application of chemicals to rice paddies, given a previous
calculation of local hydrological balance as input.
}
\details{
TODO: Document columns of returned df #72
}
