% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca--main.R
\name{ca}
\alias{ca}
\alias{compute_ca}
\title{CA: Chemical Applications}
\usage{
ca(model)

compute_ca(model, ca_schedules_df = erahumed::albufera_ca_schedules)
}
\arguments{
\item{model}{An object of class \link{erahumed_model}, with a pre-computed
\link{hbp} component (\emph{i.e.} such that \code{hbp(model)} is not \code{NULL}).}

\item{ca_schedules_df}{a \code{data.frame} following the template of
\link{albufera_ca_schedules}. Each row of this data.frame corresponds to a
scheduled application. The semantics of columns are the same as in
\link{albufera_ca_schedules}.}
}
\value{
Objects of class \link{erahumed_model} and \code{erahumed_ca}, for
\code{compute_ca()} and \code{ca()} respectively.
}
\description{
This model component simulates the application of chemicals to rice paddy
clusters, based on a previously computed simulation of local hydrological
balance. The result is a set of time series of applied doses, one for each
applied chemical.

This modeling layer requires the \link{hbp} component of the model to be
pre-computed.
}
\details{
The output \code{data.frame} extends the output of the underlying \link{hbp}
layer, preserving its cardinality (one row per cluster per day). The
additional columns, named as the chemicals appearing in \code{ca_schedules_df},
provide the time series of applied doses, expressed in kilograms.
}
\seealso{
Other model components: 
\code{\link{ct}()},
\code{\link{hba}()},
\code{\link{hbp}()},
\code{\link{inp}()}
}
\author{
Pablo Amador Crespo, Valerio Gherardi
}
\concept{model components}
