% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta--component-output.R
\name{component_output}
\alias{component_output}
\alias{component_output.erahumed_model_component}
\alias{component_output.erahumed_model}
\title{Extract model component outputs}
\usage{
component_output(object, ...)

\method{component_output}{erahumed_model_component}(object, ...)

\method{component_output}{erahumed_model}(object, component, ...)
}
\arguments{
\item{object}{The object from which the component output is to be extracted
from (see above for available methods).}

\item{...}{further arguments passed to or from other methods.}

\item{component}{A string. Name of the component whose output \code{data.frame}
is to be extracted.}
}
\value{
A \code{data.frame}.
}
\description{
Extracts the output \code{data.frame} of a \link{erahumed_model_component}.
}
\author{
Valerio Gherardi
}
