% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_balance.R
\name{compute_ditch_inflow_pct}
\alias{compute_ditch_inflow_pct}
\title{Compute Ditch Inflow Percent}
\usage{
compute_ditch_inflow_pct(cluster_data = clusters)
}
\arguments{
\item{clusters_data}{A dataframe containing two columns \code{ditch} and \code{area}.
Each row is assumed to represent a distinct cluster pertaining to \code{ditch},
and with the surface specified by \code{area}.}
}
\value{
A dataframe containing two columns, \code{ditch} and \code{inflow_pct},
specifying a ditch and the amount of inflow corresponding to it,
respectively.
}
\description{
Computes the fraction of water that flows through each ditch, as
the fraction of total surface covered by the clusters adjacent to said ditch.
}
\details{
TODO: other options to obtain ditch percents (e.g. empirical data by Soria et al.)?
}
\author{
Pablo Amador Crespo, Valerio Gherardi
}
