% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct--main.R
\name{ct}
\alias{ct}
\alias{compute_ct}
\title{CT: Chemical Transport}
\usage{
ct(model)

compute_ct(model)
}
\arguments{
\item{model}{An object of class \link{erahumed_model}, with a pre-computed
\link{ca} component (\emph{i.e.} such that \code{ca(model)} is not \code{NULL}).}
}
\value{
Objects of class \link{erahumed_model} and \code{erahumed_ct}, for
\code{compute_ct()} and \code{ct()} respectively.
}
\description{
This model component computes the evolution of chemicals applied to rice
paddy clusters, based on the previously computed simulations for
hydrological balance and chemicals application.
The result is a set of time series of concentrations, one for each
applied chemical.

This modeling layer requires the \link{ca} component of the model to be
pre-computed.
}
\details{
TBD.
}
\seealso{
Other model components: 
\code{\link{ca}()},
\code{\link{hba}()},
\code{\link{hbp}()},
\code{\link{inp}()}
}
\author{
Pablo Amador Crespo, Valerio Gherardi
}
\concept{model components}
