% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ditch_inflow_pcts}
\alias{ditch_inflow_pcts}
\title{Ditch Inflow Percents}
\format{
\subsection{\code{ditch_inflow_pcts}}{

A data frame with 6140 rows and 5 columns:
\describe{
\item{mes}{Character. Month name (in Spanish).}
\item{acqX}{Double. Percent of total inflow contributed by ditch number X.}
}
}
}
\source{
\url{https://www.mdpi.com/2306-5338/8/1/37}
}
\usage{
ditch_inflow_pcts
}
\description{
Percents of total inflows separated by ditch, month by month. Based on
observational data.
}
\keyword{datasets}
