% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs--erahumed-package.R
\docType{package}
\name{erahumed-package}
\alias{erahumed}
\alias{erahumed-package}
\title{erahumed: ERAHUMED Data Analysis and Visualization}
\description{
ERAHUMED Data Analysis and Visualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/erahumed/erahumed}
  \item \url{https://erahumed.github.io/erahumed/}
  \item Report bugs at \url{https://github.com/erahumed/erahumed/issues}
}

}
\author{
\strong{Maintainer}: Valerio Gherardi \email{vgherard840@gmail.com} (\href{https://orcid.org/0000-0002-8215-3013}{ORCID})

Authors:
\itemize{
  \item Pablo Amador \email{pamadorc@gmail.com} (\href{https://orcid.org/0000-0002-3793-8794}{ORCID})
}

Other contributors:
\itemize{
  \item Andreu Rico (\href{https://orcid.org/0000-0002-1820-4218}{ORCID}) [reviewer]
}

}
\keyword{internal}
