% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta--layers.R
\name{erahumed_layers}
\alias{erahumed_layers}
\title{ERAHUMED simulation layers}
\usage{
erahumed_layers()
}
\value{
A character vector. List of ERAHUMED simulation layers.
}
\description{
Returns a list of ERAHUMED simulation layers
(\emph{cf.} \link{erahumed_simulation_interface}).
}
\details{
For further information on the various layers, see \verb{?<layer_name>}.
}
