% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta--simulation.R
\name{erahumed_simulation}
\alias{erahumed_simulation}
\title{Initialize an ERAHUMED simulation}
\usage{
erahumed_simulation()
}
\value{
An object of class \code{erahumed_simulation}.
}
\description{
Initializes an ERAHUMED simulation. See \link{erahumed_simulation_interface}
for a detailed description of the simulation workflow.
}
\examples{
erahumed_simulation()

}
\author{
Valerio Gherardi
}
