% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta--simulation.R
\name{erahumed_simulation}
\alias{erahumed_simulation}
\title{Initialize an ERAHUMED simulation}
\usage{
erahumed_simulation()
}
\value{
An object of class \code{erahumed_simulation}, that represents a blank ERAHUMED
simulation (with no simulation layer computed yet).
}
\description{
Initializes an ERAHUMED simulation. This is the starting point for the ERAHUMED
simulation chain, whose computations are performed by \verb{compute_*()} functions.
See \link{erahumed_simulation_interface} for more details.
}
\examples{
m <- erahumed_simulation()
m

}
\author{
Valerio Gherardi
}
