% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta--simulation.R
\name{erahumed_simulation}
\alias{erahumed_simulation}
\title{Initialize an ERAHUMED simulation}
\usage{
erahumed_simulation()
}
\value{
An object of class \code{erahumed_simulation}.
}
\description{
Initializes an ERAHUMED simulation. Check the
\href{https://erahumed.github.io/erahumed/articles/erahumed-workflow.html}{main package vignette}
for a detailed description of the simulation workflow.
}
\examples{
erahumed_simulation()

}
