% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta--layer-extractors.R
\name{get_layer}
\alias{get_layer}
\title{Extract simulation layers}
\usage{
get_layer(simulation, layer)
}
\arguments{
\item{simulation}{The simulation from which the layer is to be extracted
from.}

\item{layer}{A string. Name of the layer to be extracted.}
}
\value{
A \link{erahumed_simulation_layer}.
}
\description{
Extracts an \link{erahumed_simulation_layer}.
}
\details{
At present, the main purpose of extracting full layer objects is to use
the predefined plotting methods of \code{{erahumed}} (see example below).
}
\examples{
\dontrun{
# Generate a dygraphs plot of the volume time-series
erahumed_simulation() |>
  run_simulation("hbl") |>
  get_layer("hbl") |>
  plot("volume")
}
}
