% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hba.R
\name{hba}
\alias{hba}
\alias{compute_hba}
\title{HBA: Hydrological Balance of the Albufera Lake}
\usage{
hba(model)

compute_hba(
  model,
  storage_curve = linear_storage_curve(intercept = 16.7459 * 1e+06, slope = 23.6577 *
    1e+06),
  petp_surface = linear_petp_surface(surface_P = 114.225826072 * 1e+06, surface_ETP =
    79.360993685 * 1e+06)
)
}
\arguments{
\item{model}{An object of class \link{erahumed_model}, with a pre-computed
\link{inp} component (\emph{i.e.} such that \code{inp(model)} is not \code{NULL}).}

\item{storage_curve}{a function that takes a numeric vector as input, and
returns a numeric vector of the same length. Function that converts lake
levels into lake \emph{volumes}.}

\item{petp_surface}{a function that takes two numeric vectors of common
length as inputs, and returns a numeric vector of the same length. Function
that converts precipitation and evapotranspiration values into an overall
volume change.}
}
\value{
Objects of class \link{erahumed_model} and \code{erahumed_hba}, for
\code{compute_hba()} and \code{hba()} respectively. In particular, the output
data.frame of the HBA component has the following calculated columns:
\itemize{
\item \code{volume} Volume time series, obtained from the storage curve.
\item \code{volume_change} Differenced time series of volume. The \eqn{n}-th is given
by \eqn{\Delta V _n \equiv V_{n+1}-V_n}, where \eqn{V_n} is volume at time
step \eqn{n}.
\item \code{petp_change} Time series. The amount \eqn{\Delta V _n ^\text{P-ETP}} of
volume change due to precipitation and evapotranspiration, obtained from the
P-ETP surface.
\item \code{outflow_total} Time series of total outflows, measured in cube meters per
second. This is the sum
\eqn{\sum _i O_i} of time series passed through the \code{outflows} argument,
plus an extra term \eqn{\delta O} of unaccounted outflow. The correction term
is defined by
\eqn{\delta O = \frac{\Delta V _n - \Delta V _n ^\text{P-ETP}}{24 \times 60 \times 60} -\sum _i O_i}
and is chosen in such a way to ensure that the total inflow is always
non-negative.
\item \code{outflow_extra} Time series. The unaccounted outflow term \eqn{\delta O}
described above.
\item \code{inflow_total} Time serie of total inflows, in cube meters per second.
This is computed as \eqn{I = \sum _{i} O_i + \delta O + \frac{\Delta V _n - \Delta V _n ^\text{P-ETP}}{24 \times 60 \times 60}}.
\item \code{residence_time_days}. Residence time, as modeled by \link{hba_residence_time}.
}
}
\description{
This model component completes the partial hydrological balance information
provided as input data, by computing the total inflow to the Albufera lake,
as well as, if relevant, the amount of unaccounted outflow - usually
attributed to \emph{tancats} that suck water from the lake.

It is the second modeling layer, and requires the \link{inp} component of the
model to be previously defined.
}
\details{
The numeric inputs for the linear storage curve are taken from the CHJ report
\href{https://www.chj.es/Descargas/ProyectosOPH/Consulta\%20publica/PHC-2015-2021/ReferenciasBibliograficas/HumedalesZonasProtegidas/CHJ,2012.Aquatool_Albufera.pdf}{Modelo de seguimiento de l’Albufera de Valencia con AQUATOOLDMA.}.
The values used as the arguments of \code{petp_surface()} were calculated by the
package authors, and correspond to the total study area (\code{surface_P}) and
the flooded surface (\code{surface_ETP}).
}
\author{
Pablo Amador Crespo, Valerio Gherardi
}
