% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbd--main.R
\name{hbd}
\alias{hbd}
\alias{setup_hbd}
\title{Hydrological Balance of ditches}
\usage{
setup_hbd(simulation, ditch_level_m = 1)
}
\arguments{
\item{simulation}{\code{[}\link{erahumed_simulation}\verb{]} \cr
The simulation object being modified.}

\item{ditch_level_m}{\verb{[numeric(1)]} \cr Constant water depth in ditches.}
}
\value{
An object of class \link{erahumed_simulation}.
}
\description{
This layer simulates the hydrology of ditches of the Albufera lake.
}
\seealso{
Other simulation layers: 
\code{\link{ca}},
\code{\link{ct}},
\code{\link{hba}},
\code{\link{hbp}},
\code{\link{inp}}
}
\concept{simulation layers}
