% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbp.R
\name{hbp}
\alias{hbp}
\alias{compute_hbp}
\title{Hydrological Balance of rice Paddies}
\usage{
hbp(model)

compute_hbp(
  model,
  management_df = erahumed::albufera_management,
  clusters_df = erahumed::albufera_clusters,
  ideal_flow_rate_cm = 5
)
}
\arguments{
\item{outflows_df, petp_df, management_df, clusters_df}{\code{data.frame}s, whose
structures follow the templates of
\link{albufera_outflows}, \link{albufera_petp}, \link{albufera_management}
and \link{albufera_clusters}, respectively.}
}
\value{
An object of class \code{hbp}, a lightweight wrapper of \code{data.frame}
with a few additional visualization methods (most prominently
\link{plot.hbp}).
}
\description{
Wrapper around \link{hbp}, used to run the local hydrological balance
simulation algorithm with the data for the Albufera lake packed up in the
\code{data.frame}s exported by \code{erahumed}.
}
\details{
The numeric inputs used from the linear storage curve and P-ETP surface were
extracted from the CHJ report
\href{https://www.chj.es/Descargas/ProyectosOPH/Consulta\%20publica/PHC-2015-2021/ReferenciasBibliograficas/HumedalesZonasProtegidas/CHJ,2012.Aquatool_Albufera.pdf}{Modelo de seguimiento de l’Albufera de Valencia con AQUATOOLDMA.}
}
