% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbp--main.R
\name{hbp}
\alias{hbp}
\alias{setup_hbp}
\title{HBP: Hydrological Balance of rice Paddies}
\usage{
setup_hbp(
  simulation,
  management_df = erahumed::albufera_management,
  ideal_flow_rate_cm = 5,
  height_thresh_cm = 2,
  seed = 840
)
}
\arguments{
\item{simulation}{An object of class \link{erahumed_simulation}.}

\item{management_df}{A \code{data.frame} that provides the yearly schedule for
irrigation and draining, that strictly follows the template of
\link{albufera_management}.}

\item{ideal_flow_rate_cm}{A positive number. Ideal inflow/outflow of a
cluster, for days in which the cluster is scheduled to be in flux (\emph{i.e.}
when being simultaneously irrigated and drained) Expressed in centimeters
per day.}

\item{height_thresh_cm}{A positive number. Height threshold for water levels,
below which a cluster is considered to be emptied.}
}
\value{
An object of class \link{erahumed_simulation}.
}
\description{
This simulation layer performs a simulation of hydrological balance
daily data for the rice paddies of the Albufera National Park, that provides
a synthetic dataset of daily inflows, outflows and average water levels for
each rice paddy.

The inputs to this simulation are:
\enumerate{
\item The previously calculated hydrological balance for the Albufera lake.
\item A geographical subdivision of the Albufera National Park into "clusters"
of rice paddies, of a definite rice variety, and whose waters flow into the
same ditch.
\item An yearly ideal schedule for irrigation and draining, with corresponding
expected water levels of a cluster for each day of the year.
}

The data for the second input is stored internally in \code{{erahumed}}, and can
be examined through the \link{clusters} helper.
}
\details{
The output \code{data.frame} for this model layer, retrieved through
\code{get_layer_output(layer = "hbp")} has one row per cluster and per day,
providing the simulated hydrological times-series for all paddy clusters. The
data-set has the following columns:

\describe{
\item{date}{Date of measurement}
\item{cluster_id}{Cluster identifier.}
\item{area_m2}{Surface area (in squared meters) of cluster.}
\item{tancat}{\code{TRUE} or \code{FALSE}, whether the cluster is a \emph{tancat}
(\emph{cf.} \link{clusters}).}
\item{variety}{Variety of rice planted in the cluster.
(\emph{cf.} \link{clusters}).}
\item{area_m2}{Surface area (in squared meters) of cluster.}
\item{ditch}{Ditch to which the cluster pertains (\emph{cf.}
\link{clusters}).}
\item{capacity_m3_s}{Outflow (to the Albufera Lake) of ditch.}
\item{seed_day}{Whether \code{date} corresponds to the sowing day of the year.}
\item{petp_cm}{Precipitation minus evapotranspiration per unit area, in
centimeters (common to all clusters).}
\item{ideal_draining}{\code{TRUE} or \code{FALSE}, whether the cluster is supposed to
be drained on this day of the year.}
\item{ideal_irrigation}{\code{TRUE} or \code{FALSE}, whether the cluster is supposed
to be irrigated on this day of the year.}
\item{ideal_inflow_cm}{Ideal inflow computed by the HBP algorithm
(see TODO #46).}
\item{ideal_outflow_cm}{Ideal outflow computed by the HBP algorithm
(see TODO #46).}
\item{ideal_diff_flow_cm}{Ideal inflow minus ideal outflow.}
\item{ideal_height_eod_cm}{Ideal water level of cluster at the end of day
(\emph{i.e.} after draining and/or irrigation, for this day of the year.}
\item{draining}{\code{TRUE} or \code{FALSE}, whether the cluster is actually (\emph{i.e.}
in the simulation) being drained on this day of the year.}
\item{irrigation}{\code{TRUE} or \code{FALSE}, whether the cluster is actually (
\emph{i.e.} in the simulation) being irrigated on this day of the year.}
\item{height_eod_cm}{Simulated water level of cluster at the end of day.}
\item{height_sod_cm}{Simulated water level of cluster at the end of day (
\emph{i.e.} before irrigation and/or draining).}
\item{inflow_cm}{Simulated inflow of cluster (in centimeters).}
\item{inflow_m3_s}{Simulated inflow of cluster (in cube meters per
second).}
\item{outflow_cm}{Simulated outflow of cluster (in centimeters).}
\item{outflow_m3_s}{Simulated outflow of cluster (in cube meters per
second).}
\item{plan_delay}{Accumulated delay on the yearly plan for this cluster.
See (TODO #46) for a detailed description.}
}
}
\seealso{
Other simulation layers: 
\code{\link{ca}},
\code{\link{ct}},
\code{\link{hba}},
\code{\link{inp}}
}
\concept{simulation layers}
