% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hb.R
\name{hydro_balance_global}
\alias{hydro_balance_global}
\title{Global Hydrological Balance}
\usage{
hydro_balance_global(
  level,
  P,
  ETP,
  outflows,
  storage_curve = erahumed::linear_storage_curve(slope = 1, intercept = 0),
  petp_surface = erahumed::linear_petp_surface(surface_P = 1, surface_ETP = 1)
)
}
\description{
Documentation TBD
}
