% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux--gis.R
\name{hydrogeo_components}
\alias{hydrogeo_components}
\alias{info_clusters}
\alias{info_ditches}
\title{Landscape hydro-geographical components of the Albufera Natural Park}
\usage{
info_clusters(include_geometry = FALSE)

info_ditches(include_geometry = FALSE)
}
\arguments{
\item{include_geometry}{\code{TRUE} or \code{FALSE}. Whether to include the geometries
of the various elements (as a column of class \code{sfc_MULTIPOLYGON} from the
\code{{sf}} package).}
}
\value{
A \code{data.frame}.
}
\description{
These helpers return \code{data.frame}s containing information on the
hydro-geographical components used by ERAHUMED to model the Albufera
Natural Park ecosystem: ditches and rice-field clusters.
For a detailed description of how these spatial units are represented
and modelled within ERAHUMED, see the
\href{https://erahumed.github.io/erahumed-book/chapters/birdseye.html#sec-hydrogeo}{user manual}.
}
