% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inp--main.R
\name{inp}
\alias{inp}
\alias{setup_inp}
\title{INP: Input data}
\usage{
setup_inp(
  simulation,
  outflows_df = erahumed::albufera_outflows,
  weather_df = erahumed::albufera_weather,
  variety_prop = c(J.Sendra = 0.8, Bomba = 0.1, Clearfield = 0.1),
  seed = 840
)
}
\arguments{
\item{simulation}{An object of class \link{erahumed_simulation}.}

\item{outflows_df}{A \code{data.frame} that provides the observational hydrological data on the Albufera lake, along the template of \link{albufera_outflows} (the default value).}

\item{weather_df}{A \code{data.frame} that provides the relevant metereological time series, along the template of \link{albufera_weather} (the default value).}

\item{variety_prop}{Numeric vector of length \code{3} with positive entries. Controls the proportions of field surface allocated to the cultivation of the J. Sendra, Bomba, and Clearfield varieties, respectively (positive vectors that do not sum to one are allowed, with entries being understood as relative proportions).}

\item{seed}{A number. Seed for random number generation used by the simulation algorithms.}
}
\value{
An object of class \link{erahumed_simulation}.
}
\description{
This simulation layer has the only purpose to collect the observational (or,
potentially, synthetic) data used by the algorithms of the ERAHUMED decision
support system.

This is the first layer of the simulation chain and does not have
any upstream dependence.
}
\details{
Both input \code{data.frame}s should have a \code{date} column (of class \link{Date}),
and the corresponding date domain should be an interval (\emph{i.e.} no missing
data between the maximum and minimum of \code{date} is allowed). The subsequent
simulation will only be performed on the largest date interval in which both
hydrological and weather input data is available.
}
\seealso{
Other simulation layers: 
\code{\link{ca}},
\code{\link{ct}},
\code{\link{hba}},
\code{\link{hbp}}
}
\concept{simulation layers}
