% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct--plot.R
\name{plot.erahumed_ct}
\alias{plot.erahumed_ct}
\title{Plot CT simulation layer output}
\usage{
\method{plot}{erahumed_ct}(x, type = c("cluster_view"), variable = c("mass", "density"), ...)
}
\arguments{
\item{x}{An object of class \code{ct}.}

\item{type}{Type of plot to be generated. Currently, only the
\code{"cluster_view"} plot type is implemented.}

\item{variable}{Variable to plot. Either "mass" or "density".}

\item{...}{Further plotting parameters.}
}
\value{
A plotly plot.
}
\description{
Plot method for \link{ct} simulation layers.
}
\details{
The \code{"cluster_view"} plot produces a time series plot of the daily
amounts of chemical (in Kg) present in the three compartments (foliage,
water and sediment) of an individual cluster. In order to use this plotting
method, the user must provide an additional \code{cluster_id} argument, a string
specifying the identifier of the cluster whose levels are to be plotted.
}
