% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbd--plot.R
\name{plot.erahumed_hbd}
\alias{plot.erahumed_hbd}
\title{Plot HBD simulation layer output}
\usage{
\method{plot}{erahumed_hbd}(x, ditch, ...)
}
\arguments{
\item{x}{An object of class \code{hbd}.}

\item{...}{Not used}
}
\value{
A \link[dygraphs]{dygraph} plot.
}
\description{
Plot method for \link{hbd} simulation layers.
}
