% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfms--s3-class.R
\name{rfms}
\alias{rfms}
\alias{new_rfms}
\title{Rice Field Management System}
\usage{
new_rfms()
}
\value{
An object of class \code{erahumed_rfms}.
}
\description{
Functions to define and initialize rice field management systems.
}
\details{
\itemize{
\item \code{new_rfms()} creates an empty management system, without any scheduled
chemical applications.
\item \code{jsendra_rfms()}, \code{bomba_rfms()}, and \code{clearfield_rfms()} provide
predefined systems inspired by the \emph{J. Sendra}, \emph{Bomba}, and \emph{Clearfield}
rice varieties, respectively.
}

Additional chemical applications can be scheduled using the helper function
\code{\link[=add_application]{add_application()}}.
}
\seealso{
\link{add_application}
}
