% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta--run-simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Run an ERAHUMED simulation}
\usage{
run_simulation(simulation)
}
\arguments{
\item{simulation}{\code{[}\link{erahumed_simulation}\verb{]} \cr
The simulation object containing the layers to be computed.}
}
\value{
An \link{erahumed_simulation}.
}
\description{
Computes the layers of a ERAHUMED simulation. See
\link{erahumed_simulation_interface} for a detailed description of the
simulation workflow.
}
