% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfms--s3-class.R
\name{schedule_application}
\alias{schedule_application}
\title{Schedule chemical application}
\usage{
schedule_application(
  system,
  chemical,
  amount_kg_ha,
  seed_day,
  type = c("ground", "aerial"),
  emptying_days = 1
)
}
\arguments{
\item{system}{\code{[}\link[=rfms]{erahumed_rfms}\verb{]} \cr
A rice field management system object.}

\item{chemical}{\code{[}\link[=chemical]{erahumed_chemical}\verb{]} \cr
Chemical to be applied.}

\item{amount_kg_ha}{\verb{[numeric(1)]} \cr
Application rate in kilograms per hectare.}

\item{seed_day}{\verb{[integer(1)]} \cr
Day after seeding when the application occurs.}

\item{type}{\verb{[character(1)]} \cr
Application type, either \code{"ground"} or \code{"aerial"}.}

\item{emptying_days}{\verb{[numeric(1)]} \cr
Duration (in days) that the field remains empty after a \code{"ground"}
application. Ignored if \code{type} is \code{"aerial"}.}
}
\value{
An object of class \link[=rfms]{erahumed_rfms}.
}
\description{
Schedules a chemical application within an existing rice field management
system (see \link{rfms}).
}
\seealso{
\link{rfms}
}
